% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{auc}
\alias{auc}
\title{Area Under the Curve of Binomial Observations vs Probabilistic Model Predictions}
\usage{
auc(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required, binomial vector) Observations, values of a binomial response variable with unique values 0 and 1. Default: NULL}

\item{p}{(required, numeric vector) Continuous predictions of a binomial model in the range 0-1. Default: NULL}
}
\value{
numeric: Area Under the Curve
}
\description{
Internal function to compute the AUC of binomial models within \code{\link[=preference_order]{preference_order()}}. As it is build for speed, this function does not check the inputs.
}
\seealso{
Other preference_order: 
\code{\link{case_weights}()},
\code{\link{f_auc}},
\code{\link{f_r2}},
\code{\link{f_r2_counts}},
\code{\link{f_v}()},
\code{\link{f_v_rf_categorical}()},
\code{\link{preference_order}()},
\code{\link{r2}()}
}
\concept{preference_order}
