% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{auc_score}
\alias{auc_score}
\title{Area Under the Receiver Operating Characteristic}
\usage{
auc_score(observed = NULL, predicted = NULL)
}
\arguments{
\item{observed}{(required, integer) Numeric vector with observations. Valid values are 1 and 0. Must have the same length as \code{predicted}. Default: NULL}

\item{predicted}{(required, numeric) Numeric vector in the range 0-1 with binary model predictions. Must have the same length as \code{observed}.}
}
\value{
numeric: area under the curve
}
\description{
Computes the AUC score of binary model predictions.
}
\examples{

 out <- auc_score(
   observed = c(0, 0, 1, 1),
   predicted = c(0.1, 0.6, 0.4, 0.8)
   )
}
\seealso{
Other preference_order: 
\code{\link{case_weights}()},
\code{\link{f_gam_auc_balanced}()},
\code{\link{f_gam_auc_unbalanced}()},
\code{\link{f_gam_deviance}()},
\code{\link{f_logistic_auc_balanced}()},
\code{\link{f_logistic_auc_unbalanced}()},
\code{\link{f_rf_auc_balanced}()},
\code{\link{f_rf_auc_unbalanced}()},
\code{\link{f_rf_rsquared}()},
\code{\link{f_rsquared}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
