% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoconfig_arg_max_cor.R
\name{autoconfig_arg_max_cor}
\alias{autoconfig_arg_max_cor}
\title{Autoconfigure Argument \code{max_cor}}
\usage{
autoconfig_arg_max_cor(
  df = NULL,
  predictors = NULL,
  max_cor = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{max_cor}{(optional, numeric or NULL) Maximum correlation allowed between pairs of \code{predictors}.}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
numeric
}
\description{
If \code{max_cor = NULL}, computes the average pairwise correlation between all predictors as \code{x} and returns \code{max(x, 0.5)}. Otherwise, it checks \code{max_cor} with \code{\link[=validate_arg_max_cor]{validate_arg_max_cor()}} and returns the result.
}
\examples{
max_cor <- autoconfig_arg_max_cor(
  df = vi_smol,
  predictors = vi_predictors_numeric,
  max_cor = NULL
)

max_cor
}
