% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoconfig_arg_max_vif.R
\name{autoconfig_arg_max_vif}
\alias{autoconfig_arg_max_vif}
\title{Autoconfigure Argument \code{max_vif}}
\usage{
autoconfig_arg_max_vif(
  df = NULL,
  predictors = NULL,
  max_cor = NULL,
  max_vif = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{max_cor}{(optional; numeric or NULL) Maximum correlation allowed between pairs of \code{predictors}. Valid values are between 0.01 and 0.99, and recommended values are between 0.5 (strict) and 0.9 (permissive). If NULL, the pairwise correlation analysis is disabled. Default: 0.7}

\item{max_vif}{optional, numeric or NULL) Maximum Variance Inflation Factor allowed for \code{predictors} during multicollinearity filtering. Recommended values are between 2 (strict) and 10 (permissive). If NULL, the value is autoconfigured depending on the average pairwise correlation of the dataset.}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
numeric
}
\description{
If \code{max_vif = NULL}, returns the prediction of the model \link{gam_cor_to_vif} on the input \code{max_cor} (see \link{experiment_df} for further details).
}
\examples{
max_vif <- autoconfig_arg_max_vif(
  df = vi_smol,
  predictors = vi_predictors_numeric,
  max_vif = NULL
)

max_vif
}
