% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_build.collinear_output.R
\name{build.collinear_output}
\alias{build.collinear_output}
\title{Builds Class \code{collinear_output}}
\usage{
build.collinear_output(collinear_selection = NULL, collinear_arguments = NULL)
}
\arguments{
\item{collinear_selection}{(required, list) List with object/s of class \code{collinear_selection} (see \code{\link[=build.collinear_selection]{build.collinear_selection()}}). Default: NULL}

\item{collinear_arguments}{(required, list) List of class \code{collinear_arguments} (see \code{\link[=build.collinear_arguments]{build.collinear_arguments()}}). Default: NULL}
}
\value{
list:
\itemize{
\item One or several objects of class \code{collinear_selection}
\item One object of class \code{collinear arguments}
}
}
\description{
Builds the object resulting from a \code{\link[=collinear]{collinear()}} call.
}
\seealso{
Other S3_methods: 
\code{\link{build.collinear_arguments}()},
\code{\link{build.collinear_selection}()},
\code{\link{print.collinear_arguments}()},
\code{\link{print.collinear_output}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_output}()},
\code{\link{summary.collinear_selection}()}
}
\concept{S3_methods}
