% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{case_weights}
\alias{case_weights}
\title{Case Weights for Unbalanced Binary Response}
\usage{
case_weights(x = NULL)
}
\arguments{
\item{x}{(required, integer vector) Binary response with 1s and 0s. Default: \code{NULL}}
}
\value{
A vector with a length equal to \code{x} with case weights.
}
\description{
Case Weights for Unbalanced Binary Response
}
\examples{
if(interactive()){

 case_weights(
   x = c(0, 0, 0, 1, 1)
 )

 }
}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_r2_continuous}},
\code{\link{f_r2_counts}},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
