% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_class.collinear_output.R
\name{class.collinear_output}
\alias{class.collinear_output}
\title{Builds Class \code{class.collinear_output}}
\usage{
class.collinear_output(collinear_selection = NULL, collinear_arguments = NULL)
}
\arguments{
\item{collinear_selection}{(required, list) Output of  \code{\link[=class.collinear_selection]{class.collinear_selection()}}. Default: NULL}

\item{collinear_arguments}{(required, list) Output of \code{\link[=class.collinear_arguments]{class.collinear_arguments()}}. Default: NULL}
}
\value{
list:
\itemize{
\item One or several objects of class \link{class.collinear_selection}.
\item One object of class \link{class.collinear_arguments}
}
}
\description{
Builds the object resulting from a \code{\link[=collinear]{collinear()}} call.
}
\seealso{
Other S3_methods: 
\code{\link{class.collinear_arguments}()},
\code{\link{class.collinear_selection}()},
\code{\link{print.collinear_arguments}()},
\code{\link{print.collinear_output}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_output}()},
\code{\link{summary.collinear_selection}()}
}
\concept{S3_methods}
