% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_class.collinear_selection.R
\name{class.collinear_selection}
\alias{class.collinear_selection}
\title{Builds \code{class.collinear_selection}}
\usage{
class.collinear_selection(
  df = NULL,
  response = NULL,
  preference_order = NULL,
  selection = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional, character string) Name of a response variable in \code{df}. Default: NULL.}

\item{preference_order}{(optional; character vector, output of \code{\link[=preference_order]{preference_order()}}, or NULL). Incompatible with \code{f} (overrides it when provided). Prioritizes predictors to preserve the most relevant ones during multicollinearity filtering.

Accepted inputs are:

\itemize{

\item \strong{NULL} (default): If argument \code{f} is NULL (default), predictors are ranked from lower to higher multicollinearity. Otherwise, \code{\link[=preference_order]{preference_order()}} ranks the predictors according to their relationship with \code{responses} using the function defined in \code{f}. NOTE: The output of this setting might differ to an external call to \code{\link[=preference_order]{preference_order()}} if target encoding is triggered.

\item \strong{character vector}: Predictor names in a user-defined priority order. The first predictor in this vector is always selected, unless it has near zero-variance values. This option sets \code{f} to NULL.

\item \strong{data frame}: output of \code{\link[=preference_order]{preference_order()}} computed on the given \code{responses}. This option sets \code{f} to NULL.

\item \strong{named list}: list of data frames, output of \code{\link[=preference_order]{preference_order()}} when argument \code{responses} is a vector of length two or more. This option sets \code{f} to NULL.
}. Default: NULL}

\item{selection}{(required, character vector) Names of the selected predictors. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
list
TODO: describe list structure
}
\description{
Builds \code{class.collinear_selection}
}
\seealso{
Other S3_methods: 
\code{\link{class.collinear_arguments}()},
\code{\link{class.collinear_output}()},
\code{\link{print.collinear_arguments}()},
\code{\link{print.collinear_output}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_output}()},
\code{\link{summary.collinear_selection}()}
}
\concept{S3_methods}
