% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{collinear-package}
\alias{collinear}
\alias{collinear-package}
\title{collinear: Automated Multicollinearity Management}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://blasbenito.github.io/collinear/}
}

}
\author{
\strong{Maintainer}: Blas Benito \email{blasbenito@gmail.com} (\href{https://orcid.org/0000-0001-5105-7232}{ORCID})

}
