% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes.R
\name{collinear_arguments-class}
\alias{collinear_arguments-class}
\alias{collinear_arguments}
\title{Class \code{collinear_arguments}}
\description{
List with the arguments used in a \code{\link[=collinear]{collinear()}} call.
}
