% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinear_auto.R
\name{collinear_auto}
\alias{collinear_auto}
\title{Automated Multicollinearity Management}
\usage{
collinear_auto(
  df = NULL,
  responses = NULL,
  predictors = NULL,
  max_cor = NULL,
  max_vif = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{responses}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}. When \code{encoding_method} is not NULL, response/s are used as reference to map categorical predictors, if any, to numeric (see \code{\link[=target_encoding_lab]{target_encoding_lab()}}). When \code{f} is not NULL, responses are used to rank predictors and preserve important ones during multicollinearity filtering (see \code{\link[=preference_order]{preference_order()}}). If no response is provided, the predictors are ranked from lower to higher multicollinearity. When several responses are provided, the selection results are named after each response in the output list. If no response is provided, the variable selection shows with the name "result" in the output list. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{max_cor}{(optional, numeric or NULL) Maximum correlation allowed between pairs of \code{predictors}. If NULL, \code{\link[=autoconfig_arg_max_cor]{autoconfig_arg_max_cor()}} finds a suitable value. Default: NULL}

\item{max_vif}{(optional, numeric or NULL). Maximum Variance Inflation Factor allowed for \code{predictors} during multicollinearity filtering. If NULL, \code{\link[=autoconfig_arg_max_vif]{autoconfig_arg_max_vif()}} finds a suitable value for the given data. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
list of class \code{class.collinear_output}
}
\description{
A fire-and-forget wrapper to \code{\link[=collinear]{collinear()}}. It configures target encoding, preference order, and the maximum multicollinearity levels (\code{max_vif} and \code{max_cor}) automatically.

Target encoding is triggered with the method "loo" (leave-one-out, see \code{\link[=target_encoding_lab]{target_encoding_lab()}} for further details) when a response is numeric and there are categorical predictors.

Preference order is computed automatically when \code{responses} is not NULL using \code{\link[=f_auto]{f_auto()}} to select a proper method depending on the nature of the response and the predictors.

Arguments \code{max_vif} and \code{max_cor} are autoconfigured by the functions \code{\link[=autoconfig_arg_max_cor]{autoconfig_arg_max_cor()}} and \code{\link[=autoconfig_arg_max_vif]{autoconfig_arg_max_vif()}} depending on the overall pairwise correlation between predictors.
}
\examples{

data(
  vi_smol,
  vi_predictors_numeric
)

x <- collinear_auto(
  df = vi_smol,
  responses = "vi_numeric",
  predictors = vi_predictors_numeric
)

x

}
\seealso{
Other automated_multicollinearity_analysis: 
\code{\link{collinear}()},
\code{\link{collinear_lite}()}
}
\concept{automated_multicollinearity_analysis}
