% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_classes.R
\name{collinear_output-class}
\alias{collinear_output-class}
\alias{collinear_output}
\title{Class \code{collinear_output}}
\description{
List resulting from a \code{\link[=collinear]{collinear()}} call.
}
\details{
When argument \code{response} is NULL, the list has two elements:
\itemize{
\item \code{result}: list of class \link{collinear_selection} with the results of the multicollinearity filtering.
\item \code{arguments}: list of class \link{collinear_arguments} with the validated arguments used in the call. These may differ from the user input, and are provided for transparency.
}

The class has two printing methods:
\itemize{
\item \code{\link[=print.collinear_outpout]{print.collinear_outpout()}}: applies \code{\link[=print.collinear_selection]{print.collinear_selection()}} and \link{print.collinear_arguments} to the object.
\item \code{\link[=summary.collinear_output]{summary.collinear_output()}}: shows a summary of the multicollinearity filtering results via \code{\link[=summary.collinear_selection]{summary.collinear_selection()}}.
}
}
