% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{collinear_selection-class}
\alias{collinear_selection-class}
\alias{collinear_selection}
\title{Class \code{collinear_selection}
TODO: update class definition}
\description{
Output of the \code{collinear()} function when \code{response} is NULL or length 1. It has the custom print method \link{print.collinear_selection}.
}
\section{Slots}{

\describe{
\item{\code{response}}{(character or NULL): Name of the response variable.}

\item{\code{predictors}}{(character vector or NULL): Names of predictors involved in multicollinearity filtering.}

\item{\code{selection}}{(character vector or NULL): Names of selected non-collinear predictors.}

\item{\code{df}}{(data frame or NULL): Data frame with all variables in \code{response} and \code{selection}.}

\item{\code{args}}{(list): Values of arguments used in the function call: \code{encoding_method}, \code{preference_order}, \code{f}, \code{max_cor}, and \code{max_vif}.}
}}

