% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_v.R
\name{cramer_v}
\alias{cramer_v}
\title{Correlation (Cramer's V) for categorical variables}
\usage{
cramer_v(x = NULL, y = NULL, check_input = TRUE)
}
\arguments{
\item{x}{(required; character vector) Values of a categorical variable.  Default: NULL}

\item{y}{(required; character vector) Values of a categorical variable. Must have the same length as 'x'.  Default: NULL}

\item{check_input}{(required; logical) If FALSE, disables data checking for a slightly faster execution. Default: TRUE}
}
\value{
Numeric, value of Cramer's V
}
\description{
Cramer's V is an extension of the chi-squared test to measure the strength of association between two categorical variables. Provides values between 0 and 1, where 0 indicates no association, and 1 indicates a perfect association. In essence, Cramer's V assesses the co-occurrence of the categories of two variables to quantify how strongly these variables are related.
}
\examples{
if(interactive()){
data(ecoregions)
cramer_v(
  x = ecoregions$primary_productivity,
  y = ecoregions$dominant_landcover
  )
}
}
