% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_f.R
\name{default_f}
\alias{default_f}
\title{Select Function to Compute Preference Order}
\usage{
default_f(df = NULL, response = NULL, predictors = NULL)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{response}{(optional, character string) Name of a numeric (non-numerics are ignored) response variable in \code{df}, only required if there are categorical variables named in the argument \code{predictors}. Default: NULL.}

\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}
}
\value{
function
}
\description{
Internal function to select a proper f_...() function to compute preference order depending on the types of the response variable and the predictors. The selection criteria is available as a data frame generated by \code{\link[=default_f_heuristic]{default_f_heuristic()}}.
}
\examples{
TODO
}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{default_f_heuristic}()},
\code{\link{f_auc}},
\code{\link{f_r2}},
\code{\link{f_r2_counts}},
\code{\link{f_v}()},
\code{\link{f_v_rf_categorical}()},
\code{\link{preference_order}()},
\code{\link{r2}()}
}
\concept{preference_order}
