% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_f.R
\name{default_f_heuristic}
\alias{default_f_heuristic}
\title{Heuristic to Select Default Function to Compute Preference Order}
\usage{
default_f_heuristic()
}
\value{
data frame
}
\description{
Data frame with rules used by \code{\link[=default_f]{default_f()}} to select the function f_...() used to compute preference order in \code{\link[=preference_order]{preference_order()}}.
}
\examples{
default_f_heuristic()
}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{default_f}()},
\code{\link{f_auc}},
\code{\link{f_r2}},
\code{\link{f_r2_counts}},
\code{\link{f_v}()},
\code{\link{f_v_rf_categorical}()},
\code{\link{preference_order}()},
\code{\link{r2}()}
}
\concept{preference_order}
