% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{df_drop_geometry}
\alias{df_drop_geometry}
\title{Removes geometry column in sf data frames}
\usage{
df_drop_geometry(df)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}
}
\value{
The input data frame without a geometry column
}
\description{
Removes geometry column in sf data frames
}
\keyword{internal}
