% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{df_inspect}
\alias{df_inspect}
\title{Internal function to inspect the df argument}
\usage{
df_inspect(df = NULL, min_rows = 30)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}

\item{min_rows}{(required; integer) Minimum number of rows required for a pairwise correlation or a variance inflation factor analysis. Default: 30}
}
\value{
data frame
}
\description{
Performs the following actions:
\itemize{
\item Stops if 'df' is NULL.
\item Stops if 'df' cannot be coerced to data frame.
\item Removes geometry column if required.
\item Removes non-numeric columns with as many unique values as rows df has.
\item Raise warning if number of rows of 'df' is lower than 'min_rows'.
}
}
\examples{
if (interactive()) {
  data(ecoregions)
  ecoregions <- df_inspect(df = ecoregions)

}
}
\keyword{internal}
