% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment_adaptive_thresholds}
\alias{experiment_adaptive_thresholds}
\title{Testing the Adaptive Thresholds Selection Method Implemented in \code{\link[=collinear]{collinear()}}}
\format{
A dataframe with 10000 rows and 7 variables:
\describe{
\item{input_rows}{Number of rows in the input data.}
\item{input_predictors}{Number of predictors in the input data.}
\item{output_predictors}{Number of predictors selected by \code{\link[=collinear]{collinear()}}.}
\item{input_cor_median}{Median correlation of the input predictors as returned by \code{\link[=collinear_stats]{collinear_stats()}}.}
\item{output_cor_median}{Median correlation of the selected predictors.}
\item{input_cor_max}{Maximum correlation of the input predictors as returned by \code{\link[=collinear_stats]{collinear_stats()}}.}
\item{output_cor_max}{Maximum correlation of the selected predictors.}
\item{input_vif_max}{Maximum VIF of the input predictors as returned by \code{\link[=collinear_stats]{collinear_stats()}}.}
\item{output_vif_max}{Maximum VIF of the selected predictors.}
}
}
\usage{
experiment_adaptive_thresholds
}
\description{
A dataframe with 10,000 experiment to test the efficacy of the adaptive multicollinearity threshold used by \code{\link[=collinear]{collinear()}} when \code{max_cor} and \code{max_vif} are NULL.
}
\details{
The source data is \link{vi} plus 200 synthetic columns, for a total of 249 columns and 30.000 rows that are randomly subset on each iteration.
}
\examples{
data(experiment_adaptive_thresholds)
str(experiment_adaptive_thresholds)

#experiment's code
system.file(
  "experiments/validation_adaptive_thresholds.R",
  package = "collinear"
  )
}
\keyword{datasets}
