% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_cramer_v}
\alias{f_cramer_v}
\title{Cramer's V Between Discrete Responses and Predictors}
\usage{
f_cramer_v(x, y, df)
}
\arguments{
\item{df}{(required, data frame) data frame with the numeric column 'x' with a continuous predictor and a continuous (including integer counts) response.}
}
\value{
numeric: Cramer's V
}
\description{
Please see \code{\link[=cramer_v]{cramer_v()}} for details.
}
\examples{
TODO complete this example

}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_r2_pearson}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
