% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_tools.R
\name{f_default_rules}
\alias{f_default_rules}
\title{Rules to Select Default f Argument to Compute Preference Order}
\usage{
f_default_rules()
}
\value{
data frame
}
\description{
Data frame with rules used by \code{\link[=f_default]{f_default()}} to select the function \code{f} to compute preference order in \code{\link[=preference_order]{preference_order()}}.
}
\examples{
f_default_rules()
}
\seealso{
Other preference_order_tools: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_default}()},
\code{\link{r2}()}
}
\concept{preference_order_tools}
