% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_gam_deviance}
\alias{f_gam_deviance}
\title{Compute the preference order for predictors based on the GAM deviance explained.}
\usage{
f_gam_deviance(x, y, df)
}
\arguments{
\item{x}{The predictor variable name.}

\item{y}{The response variable name.}

\item{df}{The data frame containing the variables.}
}
\value{
The deviance explained by the GAM model for the predictor.
}
\description{
This function calculates the preference order of predictors based on the deviance explained by a Generalized Additive Model (GAM) between each predictor and the response variable.
}
