% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_gam_poisson_rsquared}
\alias{f_gam_poisson_rsquared}
\title{R-squared From a Univariate Poisson GAM Model}
\usage{
f_gam_poisson_rsquared(x, y, df)
}
\arguments{
\item{df}{(required, data frame) data frame with the numeric column 'x' with a continuous predictor and a continuous (including integer counts) response.}
}
\value{
R-squared
}
\description{
Computes the R-squared of a univariate Generalized Additive Model (GAM) fitted with the formula \code{formula = y ~ s(x)} and the family \code{stats::poisson(link = "log")}.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

#this example requires "mgcv" installed in the system
if(
  requireNamespace(
    package = "mgcv",
   quietly = TRUE
   )
 ){

 #simulate counts in vi$vi_mean
 vi$vi_mean_count <- as.integer(vi$vi_mean * 1000)

  f_gam_poisson_rsquared(
    x = "growing_season_length", #predictor
    y = "vi_mean_count",         #response
    df = vi
  )

}

}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_r2_pearson}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
