% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_glm_binomial_balanced_auc}
\alias{f_glm_binomial_balanced_auc}
\title{AUC of Binomial GLM with Logit Link for Balance Binary Responses}
\usage{
f_glm_binomial_balanced_auc(x, y, df)
}
\arguments{
\item{df}{(required, data frame) data frame with the numeric column 'x' with a continuous predictor and a continuous (including integer counts) response.}
}
\value{
numeric: area under the curve
}
\description{
Fits a logistic GLM model \code{y ~ x} when \code{y} is a binary response with values 0 and 1 and \code{x} is numeric. This function is suitable when the response variable is balanced. If the response is unbalanced, then \code{\link[=f_logistic_auc_unbalanced]{f_logistic_auc_unbalanced()}} should provide better results.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

f_logistic_auc_balanced(
  x = "growing_season_length", #predictor
  y = "vi_binary",             #binary response
  df = vi
)

}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_r2_pearson}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
