% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_glm_binomial_unbalanced_auc}
\alias{f_glm_binomial_unbalanced_auc}
\title{AUC of Binomial GLM with Logit Link for Unbalanced Binary Responses}
\usage{
f_glm_binomial_unbalanced_auc(x, y, df)
}
\arguments{
\item{df}{(required, data frame) with columns:
\itemize{
\item "x": (numeric) continuous predictor.
\item "y" (numeric) continous response.
}}
}
\value{
numeric: area under the curve
}
\description{
Fits a quasi-binomial GLM model \code{y ~ x} with case weights when \code{y} is an unbalanced binary response with values 0 and 1 and \code{x} is numeric. It uses the function \code{\link[=case_weights]{case_weights()}} to weight 0s and 1s according to their frequency within \code{y}.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

f_logistic_auc_unbalanced(
  x = "growing_season_length", #predictor
  y = "vi_binary",             #binary response
  df = vi
)

}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_r2_continuous}},
\code{\link{f_r2_counts}},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
