% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_lm_coef}
\alias{f_lm_coef}
\title{Compute the preference order for predictors based on absolute value of the linear regression coefficient.}
\usage{
f_lm_coef(x, y, df)
}
\arguments{
\item{x}{The predictor variable name.}

\item{y}{The response variable name.}

\item{df}{The data frame containing the variables.}
}
\value{
The absolute linear regression coefficient for the predictor.
}
\description{
This function calculates the preference order of predictors based on the linear regression coefficient between each predictor and the response variable.
}
