% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_logistic_auc_unbalanced}
\alias{f_logistic_auc_unbalanced}
\title{AUC of Binomial GLM with Logit Link for Unbalanced Binary Responses}
\usage{
f_logistic_auc_unbalanced(x, y, df)
}
\arguments{
\item{x}{(required, character string) name of the predictor variable.}

\item{y}{(required, character string) name of the response variable}

\item{df}{(required, data frame) data frame with the columns 'x' and 'y'.}
}
\value{
numeric: area under the curve
}
\description{
Fits a quasi-binomial GLM model \code{y ~ x} with case weights when \code{y} is an unbalanced binary response with values 0 and 1 and \code{x} is numeric. It uses the function \code{\link[=case_weights]{case_weights()}} to weight 0s and 1s according to their frequency within \code{y}.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

f_logistic_auc_unbalanced(
  x = "growing_season_length", #predictor
  y = "vi_binary",             #binary response
  df = vi
)

}
\seealso{
Other preference_order: 
\code{\link{auc_score}()},
\code{\link{case_weights}()},
\code{\link{f_gam_auc_balanced}()},
\code{\link{f_gam_auc_unbalanced}()},
\code{\link{f_gam_deviance}()},
\code{\link{f_logistic_auc_balanced}()},
\code{\link{f_rf_auc_balanced}()},
\code{\link{f_rf_auc_unbalanced}()},
\code{\link{f_rf_rsquared}()},
\code{\link{f_rsquared}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
