% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_r2_pearson}
\alias{f_r2_pearson}
\alias{f_r2_spearman}
\alias{f_r2_gam_gaussian}
\alias{f_r2_gam_poisson}
\alias{f_r2_glm_gaussian_poly2}
\alias{f_r2_glm_poisson_poly2}
\alias{f_r2_rpart}
\alias{f_r2_rf}
\title{Association Between Continuous Responses and Predictors}
\usage{
f_r2_pearson(df)

f_r2_spearman(df)

f_r2_gam_gaussian(df)

f_r2_gam_poisson(df)

f_r2_glm_gaussian_poly2(df)

f_r2_glm_poisson_poly2(df)

f_r2_rpart(df)

f_r2_rf(df)
}
\arguments{
\item{df}{(required, data frame) data frame with the numeric column 'x' with a continuous predictor and a continuous (including integer counts) response.}
}
\value{
numeric: R-squared
}
\description{
These functions take a data frame with two numeric columns "x" (predictor) and "y" (response), where "y" can either be continuous or integer counts.  assess their associations directly, or fit a model and assess the R-squared of the observations versus the model predictions:
\itemize{
\item \code{f_r2_pearson()}: Pearson's R-squared between a response and a predictor.
\item \code{f_r2_spearman()}: Spearman's R-squared between a response and a predictor.
\item \code{f_r2_gam_gaussian()} Pearson's R-squared between a continuous response and the predictions of a \code{\link[mgcv:gam]{mgcv::gam()}} model with formula \code{y ~ s(x)} and family \code{stats::gaussian(link = "identity")}.
\item \code{f_r2_gam_poisson()} Pearson's R-squared between a count response and the predictions of a \code{\link[mgcv:gam]{mgcv::gam()}} model with formula \code{y ~ s(x)} and family \code{stats::poisson(link = "log")}.
\item \code{f_r2_glm_gaussian_poly2()}: Pearson's R-squared between and a response and the predictions of a GLM model with formula \code{y ~ stats::poly(x, degree = 2, raw = TRUE)} and family \code{stats::gaussian(link = "identity")}.
\item \code{f_r2_glm_poisson_poly2()} Pearson's R-squared between a count response and the predictions of a GLM model with formula \code{y ~ stats::poly(x, degree = 2, raw = TRUE)} and family \code{stats::poisson(link = "log")}.
\item \code{f_r2_rpart()}: Pearson's R-squared between a continuous or count response and the predictions of a \code{\link[rpart:rpart]{rpart::rpart()}} (recursive partition trees) model.
\item \code{f_r2_rf()}: Pearson's R-squared between a continuous or count response and the predictions of a random forest model fitted with \code{\link[ranger:ranger]{ranger::ranger()}}
}
}
\examples{
#load example data
data(vi)

#reduce size to speed-up example
vi <- vi[1:1000, ]

#numeric response and predictor
#to data frame without NAs
df <- data.frame(
  y = vi[["vi_mean"]],
  x = vi[["swi_max"]]
) |>
  na.omit()

# Continuous response

#Pearson R-squared
f_r2_pearson(df = df)

#Spearman R-squared
f_r2_spearman(df = df)

#R-squared of a gaussian gam
f_r2_gam_gaussian(df = df)

#gaussian glm with second-degree polynomials
f_r2_glm_gaussian_poly2(df = df)

#recursive partition tree
f_r2_rpart(df = df)

#random forest model
f_r2_rf(df = df)

#Count response

#simulating counts in df
df$y <- as.integer(df$y * 1000)

#GLM model with second degree polynomials an Poisson family
f_r2_glm_poisson_poly2(df = df)

#GAM model with Poisson family
f_r2_gam_poisson(df = df)

#tree models manage counts without issues too
f_r2_rpart(df = df)
f_r2_rf(df = df)


}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}

Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}

Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}

Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}

Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}

Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}

Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}

Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{f_rf_binomial_unbalanced_auc}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
