% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_rf_auc_unbalanced}
\alias{f_rf_auc_unbalanced}
\title{AUC of Random Forest Model for Unbalanced Binary Responses}
\usage{
f_rf_auc_unbalanced(x, y, df)
}
\arguments{
\item{x}{(required, character string) name of the predictor variable.}

\item{y}{(required, character string) name of the response variable}

\item{df}{(required, data frame) data frame with the columns 'x' and 'y'.}
}
\value{
numeric: area under the curve
}
\description{
Computes a univariate random forest model with weighted cases via \verb{\link[ranger]\{ranger\}} and returns the Area Under the Curve on the out-of-bag data.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

#this example requires "ranger" installed in the system
if(
  requireNamespace(
    package = "ranger",
    quietly = TRUE
    )
  ){

  f_rf_auc_unbalanced(
    x = "growing_season_length", #predictor
    y = "vi_binary",               #response
    df = vi
  )

}

}
\seealso{
Other preference_order: 
\code{\link{auc_score}()},
\code{\link{case_weights}()},
\code{\link{f_gam_auc_balanced}()},
\code{\link{f_gam_auc_unbalanced}()},
\code{\link{f_gam_deviance}()},
\code{\link{f_logistic_auc_balanced}()},
\code{\link{f_logistic_auc_unbalanced}()},
\code{\link{f_rf_auc_balanced}()},
\code{\link{f_rf_rsquared}()},
\code{\link{f_rsquared}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
