% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_rf_binomial_unbalanced_auc}
\alias{f_rf_binomial_unbalanced_auc}
\title{AUC of Random Forest Model for Unbalanced Binary Responses}
\usage{
f_rf_binomial_unbalanced_auc(x, y, df)
}
\arguments{
\item{df}{(required, data frame) data frame with the numeric column 'x' with a continuous predictor and a continuous (including integer counts) response.}
}
\value{
numeric: area under the curve
}
\description{
Computes a univariate random forest model with weighted cases via \verb{\link[ranger]\{ranger\}} and returns the Area Under the Curve on the out-of-bag data.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

#this example requires "ranger" installed in the system
if(
  requireNamespace(
    package = "ranger",
    quietly = TRUE
    )
  ){

  f_rf_auc_unbalanced(
    x = "growing_season_length", #predictor
    y = "vi_binary",               #response
    df = vi
  )

}

}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_cramer_v}()},
\code{\link{f_gam_binomial_balanced_auc}()},
\code{\link{f_gam_binomial_unbalanced_auc}()},
\code{\link{f_gam_poisson_rsquared}()},
\code{\link{f_glm_binomial_balanced_auc}()},
\code{\link{f_glm_binomial_unbalanced_auc}()},
\code{\link{f_r2_pearson}()},
\code{\link{f_rf_binomial_balanced_auc}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
