% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_rsquared}
\alias{f_rsquared}
\title{R-squared between a response and a predictor}
\usage{
f_rsquared(x, y, df)
}
\arguments{
\item{x}{(required, character string) name of the predictor variable.}

\item{y}{(required, character string) name of the response variable}

\item{df}{(required, data frame) data frame with the columns 'x' and 'y'.}
}
\value{
R-squared
}
\description{
R-squared between a response and a predictor. Fastest option to compute preference order.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

f_rsquared(
  x = "growing_season_length", #predictor
  y = "vi_mean",               #response
  df = vi
)

}
\seealso{
Other preference_order: 
\code{\link{auc_score}()},
\code{\link{case_weights}()},
\code{\link{f_gam_auc_balanced}()},
\code{\link{f_gam_auc_unbalanced}()},
\code{\link{f_gam_deviance}()},
\code{\link{f_logistic_auc_balanced}()},
\code{\link{f_logistic_auc_unbalanced}()},
\code{\link{f_rf_auc_balanced}()},
\code{\link{f_rf_auc_unbalanced}()},
\code{\link{f_rf_rsquared}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
