% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_logical_variables.R
\name{identify_logical_variables}
\alias{identify_logical_variables}
\title{Identify Valid Logical Variables}
\usage{
identify_logical_variables(
  df = NULL,
  responses = NULL,
  predictors = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{(required; dataframe, tibble, or sf) A dataframe with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{responses}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}. Default: NULL.}

\item{predictors}{(required, character vector) Names of the predictors to identify. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Used to pass internal arguments such as \code{function_name} for \link{validate_arg_function_name} in nested calls, or \code{m}, a correlation matrix generated via \code{\link[stats:cor]{stats::cor()}} or \code{\link[=cor_matrix]{cor_matrix()}}, and the cross-validation arguments of \code{\link[=preference_order]{preference_order()}}: \code{cv_iterations}, and \code{cv_training_fraction}.}
}
\value{
list:
\itemize{
\item \code{valid}: character vector with valid logical predictor names.
\item \code{invalid}: character vector with invalid logical predictor names.
}
}
\description{
Identifies logical predictors and excludes those with constant values.
}
\examples{

data(vi_smol, vi_predictors)

#invalid logical
vi_smol$logical_invalid <- TRUE

#valid logical
vi_smol$logical_valid <- sample(
  x = c(TRUE, FALSE),
  size = nrow(vi_smol),
  replace = TRUE
)

x <- identify_logical_variables(
  df = vi_smol,
  predictors = c(
    vi_predictors,
    "logical_invalid",
    "logical_valid"
  )
)

x$valid
x$invalid

}
\seealso{
Other data_types: 
\code{\link{identify_categorical_variables}()},
\code{\link{identify_numeric_variables}()},
\code{\link{identify_response_type}()},
\code{\link{identify_valid_variables}()},
\code{\link{identify_zero_variance_variables}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_types}
