% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_predictors.R
\name{identify_predictors_type}
\alias{identify_predictors_type}
\title{Identify Predictor Types}
\usage{
identify_predictors_type(df = NULL, predictors = NULL, quiet = FALSE, ...)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
character string: predictors type
}
\description{
Internal function to identify predictor types. The supported types are:
\itemize{
\item "numeric": all predictors belong to the classes "numeric" and/or "integer".
\item "categorical": all predictors belong to the classes "character" and/or "factor".
\item "mixed": predictors are of types "numeric" and "categorical".
\item "unknown": predictors of unknown type.
}
}
\examples{
data(vi_smol, vi_predictors)

identify_predictors_type(
  df = vi_smol,
  predictors = vi_predictors
)

identify_predictors_type(
  df = vi_smol,
  predictors = vi_predictors_numeric
)

identify_predictors_type(
  df = vi_smol,
  predictors = vi_predictors_categorical
)

}
\seealso{
Other data_types: 
\code{\link{identify_predictors}()},
\code{\link{identify_predictors_categorical}()},
\code{\link{identify_predictors_logical}()},
\code{\link{identify_predictors_numeric}()},
\code{\link{identify_predictors_zero_variance}()},
\code{\link{identify_response_type}()}
}
\concept{data_types}
