% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_zero_variance_predictors}
\alias{identify_zero_variance_predictors}
\title{Identify zero and near-zero-variance predictors}
\usage{
identify_zero_variance_predictors(df = NULL, predictors = NULL, decimals = 4)
}
\arguments{
\item{df}{(required; data frame or tibble) A validated data frame with numeric and/or character predictors predictors, and optionally, a response variable. Default: NULL.}

\item{predictors}{(optional; character vector) A validated vector with predictor names in 'df'. If omitted, all columns of 'df' are used as predictors. Default:'NULL'}

\item{decimals}{(required, integer) number of decimal places for the zero variance test. Default: 4}
}
\value{
character vector with names of zero and near-zero variance columns.
}
\description{
Predictors a variance of zero or near zero are highly problematic for multicollinearity analysis and modelling in general. This function identifies these predictors with a level of sensitivity defined by the 'decimals' argument. Smaller number of decimals increase the number of variables detected as near zero variance. Recommended values will depend on the range of the numeric variables in 'df'.
}
\examples{
if (interactive()) {
  #load example data
  #and its predictors
  data(
    vi,
    vi_predictors
    )

  #validate the vi data frame
  vi <- collinear::validate_df(df = vi)

  #validate predictors
  predictors <- collinear::validate_df_predictors(
  df = vi,
  predictors = vi_predictors
  )

  #get numeric predictors
  predictors.numeric <- collinear::identify_non_numeric_predictors(
    df = vi,
    predictors = predictors
    )
}
}
