% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{predictors_inspect}
\alias{predictors_inspect}
\alias{check_response_name}
\alias{non_numeric_columns}
\alias{zero_variance_columns}
\alias{numeric_columns}
\title{Checks 'predictors' argument}
\usage{
predictors_inspect(
  predictors = NULL,
  data = NULL,
  is.required = TRUE,
  numeric.only = TRUE,
  na.allowed = FALSE,
  zero.variance.allowed = FALSE,
  decimal_places = 4,
  verbose = TRUE
)

check_response_name(
  response = NULL,
  data = NULL,
  is.required = TRUE,
  na.allowed = FALSE,
  zero.variance.allowed = FALSE,
  decimal_places = 4,
  verbose = TRUE
)

non_numeric_columns(data, columns = NULL)

zero_variance_columns(data, columns = NULL, decimal_places = 4)

numeric_columns(data = NULL, columns = NULL)
}
\arguments{
\item{predictors}{predictors.argument}

\item{data}{A data frame to extract numeric columns from.}

\item{is.required}{logical}

\item{numeric.only}{logical}

\item{na.allowed}{logical, changes the check depending on whether NAs are allowed in data or not.}

\item{zero.variance.allowed}{logical}

\item{decimal_places}{Integer, number of decimal places to round \code{columns} to. Defines the tolerance of the test. Default: 4}

\item{verbose}{logical}

\item{response}{response}

\item{columns}{A character vector specifying the columns of the data frame to extract.}
}
\value{
predictor.names

response

A character vector with the names of the numeric columns in the selected data frame.

A character vector with the names of the numeric columns in the selected data frame.

A character vector with the names of the non-numeric columns in the selected data frame.
}
\description{
This function takes a data frame and a set of columns and returns the names of the numeric columns in the selected data frame.

This function takes a data frame and a set of columns and returns the names of the columns with zero variance.

This function takes a data frame and a set of columns and returns the names of the non-numeric columns in the selected data frame.
}
\keyword{internal}
