% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_tools.R
\name{k_auto}
\alias{k_auto}
\title{Auto Computation of K in GAM Smooths}
\usage{
k_auto(df = NULL, predictor = "x")
}
\arguments{
\item{df}{(required, data frame) with columns:
\itemize{
\item "x": (numeric) continuous predictor.
\item "y" (numeric) continuous response.
}}

\item{predictor}{(optiona, string) Name of a numeric column in \code{df}, to be used as predictor in a GAM model. Default: "x".}
}
\value{
integer; k value
}
\description{
Auto Computation of K in GAM Smooths
}
\examples{
k_auto(
  df = vi,
  predictor = "swi_mean"
)
}
\seealso{
Other preference_order_tools: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{r2}()}
}
\concept{preference_order_tools}
