% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_cor.R
\name{mc_cor}
\alias{mc_cor}
\title{Pairwise Pearson correlation}
\usage{
mc_cor(data = NULL, predictors = NULL, verbose = TRUE)
}
\arguments{
\item{data}{(required; data frame or tibble) A data frame with predictors. Default: \code{NULL}.}

\item{predictors}{(optional; character vector) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{data}. Default: \code{NULL}}

\item{verbose}{(optional, logical) Set to \code{FALSE} to silence messages. Default: \code{TRUE}}
}
\value{
Data frame with the columns "a" and "b" containing the names of pairs of predictors, and "cor" containing Pearson correlation between each pair of predictors.
}
\description{
Returns the pairwise Pearson correlation between all pairs of predictors in a training dataset.

Please note that near-zero variance columns are ignored by this function. Use \code{\link[=mc_auto_vif]{mc_auto_vif()}} to remove them.
}
\examples{
if(interactive()){

data(
  ecoregions,
  ecoregions_continuous_response,
  ecoregions_all_predictors
)

df <- mc_cor(
      data = ecoregions,
      predictors = ecoregions_all_predictors
)

}
}
