% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_encoder.R
\name{mean_encoder}
\alias{mean_encoder}
\title{Target encode a categorical variable based on a numeric variable.}
\usage{
mean_encoder(response = NULL, predictor = NULL, check_input = TRUE)
}
\arguments{
\item{response}{(required; numeric vector) A numeric vector generally representing a response variable. Default: NULL}

\item{predictor}{(required; character vector) A character vector of the same length as 'response', representing the categorical variable.  Default: NULL}

\item{check_input}{(required; logical) If FALSE, disables data checking for a slightly better performance when many tests are performed. Default: TRUE}
}
\value{
A numeric vector containing target-encoded values.
}
\description{
Transforms categorical variables into numeric by assigning to each category the mean of a numeric variable (usually the model response) across the group's cases.
}
\details{
This function is used by default in \code{\link[=collinear]{collinear()}}, \code{\link[=cor_select]{cor_select()}} and \code{\link[=vif_select]{vif_select()}}. If you wish to have more control over the target encoding process, you can use \code{\link[=target_encoding_lab]{target_encoding_lab()}}.
}
\examples{
if(interactive()){

#load example data frame
data(vi)

biogeo_biome_encoded <- mean_encoder(
  response = vi$vi_mean,
  predictor = vi$biogeo_biome
  )

}

}
