% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_formula.R
\name{model_formula}
\alias{model_formula}
\title{Generate Model Formulas}
\usage{
model_formula(
  df = NULL,
  response = NULL,
  predictors = NULL,
  term_f = NULL,
  term_args = NULL,
  random_effects = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional, character string) Name of a response variable in \code{df}. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{term_f}{(optional; string). Name of function to apply to each term in the formula, such as "s" for \code{\link[mgcv:s]{mgcv::s()}} or any other smoothing function, "poly" for \code{\link[stats:poly]{stats::poly()}}. Default: NULL}

\item{term_args}{(optional; string). Arguments of the function applied to each term. For example, for "poly" it can be "degree = 2, raw = TRUE". Default: NULL}

\item{random_effects}{(optional, string or character vector). Names of variables to be used as random effects. Each element is added to the final formula as \code{+(1 | random_effect_name)}. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
list if \code{predictors} is a list or length of \code{response} is higher than one, and character vector otherwise.
}
\description{
Generate Model Formulas
}
\examples{

data(vi_smol, vi_predictors_numeric)

#additive formula
x <- model_formula(
  df = vi_smol,
  response = "vi_numeric",
  predictors = vi_predictors_numeric[1:3]
)

x

#using a formula in a model
m <- stats::lm(
 formula = x,
 data = vi_smol
 )

summary(m)


#polynomial formula
x <- model_formula(
  df = vi_smol,
  response = "vi_numeric",
  predictors = vi_predictors_numeric[1:3],
  term_f = "poly",
  term_args = "degree = 3, raw = TRUE"
)

x

#gam formula
x <- model_formula(
  df = vi_smol,
  response = "vi_numeric",
  predictors = vi_predictors_numeric[1:3],
  term_f = "s"
)

x

#random effect
x <- model_formula(
  df = vi_smol,
  response = "vi_numeric",
  predictors = vi_predictors_numeric[1:3],
  random_effects = "country_name" #from vi_smol$country_name
)

x
}
\seealso{
Other modelling_tools: 
\code{\link{case_weights}()},
\code{\link{performance_score_auc}()},
\code{\link{performance_score_r2}()},
\code{\link{performance_score_v}()}
}
\concept{modelling_tools}
