% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_score_r2.R
\name{performance_score_r2}
\alias{performance_score_r2}
\title{Pearson's R-squared of Observations vs Predictions}
\usage{
performance_score_r2(o = NULL, p = NULL, ...)
}
\arguments{
\item{o}{(required, numeric vector) Observations. Default: NULL}

\item{p}{(required, numeric vector) Predictions. Default: NULL}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
numeric: Pearson R-squared
}
\description{
Internal function to compute the R-squared of observations versus predictions via \code{\link[stats:cor]{stats::cor()}}. Used within \code{\link[=f_r2_glm_gaussian]{f_r2_glm_gaussian()}}, \code{\link[=f_r2_glm_gaussian_poly2]{f_r2_glm_gaussian_poly2()}}, \code{\link[=f_r2_gam_gaussian]{f_r2_gam_gaussian()}}, \code{\link[=f_r2_rpart]{f_r2_rpart()}}, \code{\link[=f_r2_rf]{f_r2_rf()}}, \code{\link[=f_r2_glm_poisson]{f_r2_glm_poisson()}}, \code{\link[=f_r2_glm_poisson_poly2]{f_r2_glm_poisson_poly2()}}, and \code{\link[=f_r2_gam_poisson]{f_r2_gam_poisson()}}.
}
\examples{
  performance_score_r2(
    o = c(1, 1, 1, 0.5, 0.5, 0, 0),
    p = c(1, 0.8, 0.7, 0.6, 0.5, 0.1, 0)
  )
}
\seealso{
Other modelling_tools: 
\code{\link{case_weights}()},
\code{\link{model_formula}()},
\code{\link{performance_score_auc}()},
\code{\link{performance_score_v}()}
}
\concept{modelling_tools}
