% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_score_v.R
\name{performance_score_v}
\alias{performance_score_v}
\title{Cramer's V of Observations vs Predictions}
\usage{
performance_score_v(o = NULL, p = NULL, ...)
}
\arguments{
\item{o}{(required; character vector) categorical observations. Default: NULL}

\item{p}{(required; character vector) categorical predictions. Default: NULL}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
numeric: Cramer's V
}
\description{
Internal function to compute the Cramer's V of categorical observations versus categorical model predictions. Please read the help file of \code{\link[=cor_cramer_v]{cor_cramer_v()}} for further details.
}
\examples{
performance_score_v(
 o = c("a", "a", "b", "c", "c"),
 p = c("a", "b", "b", "c", "c")
 )

}
\seealso{
Other modelling_tools: 
\code{\link{case_weights}()},
\code{\link{model_formula}()},
\code{\link{performance_score_auc}()},
\code{\link{performance_score_r2}()}
}
\concept{modelling_tools}
