% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{predictors_character}
\alias{predictors_character}
\title{Extract non-numeric predictors from a data frame}
\usage{
predictors_character(df = NULL, predictors = NULL)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}

\item{predictors}{(optional, character vector) names of predictors in arguemnt 'df'. Default: NULL}
}
\value{
character vector with names of non-numeric columns.
}
\description{
Extract non-numeric predictors from a data frame
}
\keyword{internal}
