% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{predictors_inspect}
\alias{predictors_inspect}
\title{Internal function to check the 'predictors' argument}
\usage{
predictors_inspect(
  df = NULL,
  predictors = NULL,
  min_numerics = 1,
  decimals = 4
)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}

\item{predictors}{(optional; character vector) names of predictors in arguemnt 'df'. Default: NULL}

\item{min_numerics}{(required, integer) Minimum number of numeric predictors required. Default: 1}

\item{decimals}{(required, integer) number of decimal places for the zero variance test. Smaller numbers will increase the number of variables detected as near-zero variance. Recommended values will depend on the range of the numeric variables in 'df'. Default: 4}
}
\value{
predictor names
}
\description{
Performs the following actions:
\itemize{
\item Stops if 'df' is NULL.
\item If 'predictors' is NULL, uses column names of 'df' as 'predictors'in df data frames.
\item Removes column "geometry" (in sf data frames).
\item Raise warning if there are names in 'predictors' not in the column names of 'df'.
\item Stops if number of numeric columns in 'predictors' is smaller than 'min_numerics'.
\item Raise warning if there are zero-variance columns in 'predictors' and ignores them.
}
}
\keyword{internal}
