% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{predictors_numeric}
\alias{predictors_numeric}
\title{Return names of non-numeric predictors}
\usage{
predictors_numeric(df = NULL, predictors = NULL)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}

\item{predictors}{(optional, character vector) names of predictors in 'df'. Default: NULL}
}
\value{
character vector with names of non-numeric predictors.
}
\description{
Return names of non-numeric predictors
}
\keyword{internal}
