% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{predictors_zero_variance}
\alias{predictors_zero_variance}
\title{Extract zero-variance predictors from a data frame}
\usage{
predictors_zero_variance(df = NULL, predictors = NULL, decimals = 4)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}

\item{predictors}{(optional, character vector) names of predictors in arguemnt 'df'. Default: NULL}

\item{decimals}{(required, integer) number of decimal places for the zero variance test. Smaller numbers will increase the number of variables detected as near-zero variance. Recommended values will depend on the range of the numeric variables in 'df'. Default: 4}
}
\value{
character vector with names of zero-variance columns.
}
\description{
Extract zero-variance predictors from a data frame
}
\keyword{internal}
