% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_score_auc.R
\name{preference_score_auc}
\alias{preference_score_auc}
\title{Area Under the Curve of Binomial Observations vs Probabilistic Model Predictions}
\usage{
preference_score_auc(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required, binomial vector) Binomial response with values 0 and 1. Default: NULL}

\item{p}{(required, numeric vector) Predictions of binomial model. Default: NULL}
}
\value{
numeric: Area Under the ROC Curve
}
\description{
Internal function to compute the AUC of binomial models within \code{\link[=preference_order]{preference_order()}}. As it is build for speed, this function does not check the inputs.
}
\seealso{
Other preference_order_tools: 
\code{\link{case_weights}()},
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{f_functions}()},
\code{\link{preference_order_collinear}()},
\code{\link{preference_score_r2}()},
\code{\link{preference_score_v}()}
}
\concept{preference_order_tools}
