% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_score_r2.R
\name{preference_score_r2}
\alias{preference_score_r2}
\title{Pearson's R-squared of Observations vs Predictions}
\usage{
preference_score_r2(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required, numeric vector) Response values. Default: NULL}

\item{p}{(required, numeric vector) Model predictions. Default: NULL}
}
\value{
numeric: Pearson R-squared
}
\description{
Internal function to compute the R-squared of observations versus model predictions.
}
\seealso{
Other preference_order_tools: 
\code{\link{case_weights}()},
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{f_functions}()},
\code{\link{preference_order_collinear}()},
\code{\link{preference_score_auc}()},
\code{\link{preference_score_v}()}
}
\concept{preference_order_tools}
