% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_score_v.R
\name{preference_score_v}
\alias{preference_score_v}
\title{Cramer's V of Observations vs Predictions}
\usage{
preference_score_v(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required, numeric vector) Response values. Default: NULL}

\item{p}{(required, numeric vector) Model predictions. Default: NULL}
}
\value{
numeric: Cramer's V
}
\description{
Internal function to compute the Cramer's V of categorical observations versus categorical model predictions.
}
\seealso{
Other preference_order_tools: 
\code{\link{case_weights}()},
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{f_functions}()},
\code{\link{preference_order_collinear}()},
\code{\link{preference_score_auc}()},
\code{\link{preference_score_r2}()}
}
\concept{preference_order_tools}
