% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.collinear_output}
\alias{print.collinear_output}
\title{Print Method for Class \code{collinear_output}}
\usage{
\method{print}{collinear_output}(x = NULL, n = 5, ...)
}
\arguments{
\item{x}{(required, list) Object of class \code{collinear_output} resulting from \code{\link[=collinear]{collinear()}}. Default: NULL}

\item{n}{(optional, integer) Maximum number of vector elements to print. Defaults: 5.}

\item{...}{(optional) Additional arguments (currently ignored).}
}
\description{
Prints a summary of objects of the class \link{collinear_output} produced by \code{\link[=collinear]{collinear()}} when the argument \code{response} is \code{NULL} or has length one.
}
