% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{r2}
\alias{r2}
\title{Pearson's R-squared of Observations vs Predictions}
\usage{
r2(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required, numeric vector) Observations, values of a response variable. Default: NULL}

\item{p}{(required, numeric vector) Model predictions. Default: NULL}
}
\value{
numeric: Pearson R-squared
}
\description{
Internal function to compute the R-squared of observations versus model predictions.
}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{default_f}()},
\code{\link{default_f_heuristic}()},
\code{\link{f_auc}},
\code{\link{f_r2}},
\code{\link{f_r2_counts}},
\code{\link{f_v}()},
\code{\link{f_v_rf_categorical}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
