% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{rescale_vector}
\alias{rescale_vector}
\title{Rescales a numeric vector into a new range}
\usage{
rescale_vector(
  x = NULL,
  new_min = 0,
  new_max = 1,
  old_min = NULL,
  old_max = NULL
)
}
\arguments{
\item{x}{(required, numeric vector) Numeric vector. Default: \code{NULL}}

\item{new_min}{(optional, numeric) New minimum value. Default: \code{0}}

\item{new_max}{(optional_numeric) New maximum value. Default: \code{1}}

\item{old_min}{(optional, numeric) Old minimum value. Default: \code{NULL}}

\item{old_max}{(optional_numeric) Old maximum value. Default: \code{NULL}}
}
\value{
A numeric vector of the same length as x, but with its values rescaled between \code{new_min} and \code{new_max.}
}
\description{
Rescales a numeric vector to a new range.
}
\examples{
if(interactive()){

 out <- rescale_vector(
   x = rnorm(100),
   new_min = 0,
   new_max = 100,
   integer = TRUE
   )
   out

}
}
