% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{response_inspect}
\alias{response_inspect}
\title{Checks the 'response' argument}
\usage{
response_inspect(df = NULL, response = NULL, decimals = 4)
}
\arguments{
\item{df}{(required; data frame or matrix) input data frame. Default: NULL}

\item{response}{(optional, character string) Name of a numeric response variable. Character response variables are ignored. Default: NULL.}

\item{decimals}{(required, integer) number of decimal places to round \code{predictors} to. Defines the tolerance of the zero-variance test. Default: 4}
}
\value{
character string with name of the response
}
\description{
Checks the 'response' argument
}
\keyword{internal}
