% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_auc.R
\name{score_auc}
\alias{score_auc}
\title{Area Under the Curve}
\usage{
score_auc(o = NULL, p = NULL, ...)
}
\arguments{
\item{o}{(required, numeric vector) Binomial observations (values 0 and 1). Default: NULL}

\item{p}{(required, numeric vector) Prediction of binomial model in the range 0-1. Default: NULL}

\item{...}{(optional) Used to pass internal arguments such as \code{function_name} for \link{validate_arg_function_name} in nested calls, or \code{m}, a correlation matrix generated via \code{\link[stats:cor]{stats::cor()}} or \code{\link[=cor_matrix]{cor_matrix()}}, and the cross-validation arguments of \code{\link[=preference_order]{preference_order()}}: \code{cv_iterations}, and \code{cv_training_fraction}.}
}
\value{
numeric: Area Under the ROC Curve
}
\description{
Internal function to compute the AUC of binomial models within \code{\link[=preference_order]{preference_order()}}. Used within \code{\link[=f_binomial_glm]{f_binomial_glm()}}, \code{\link[=f_binomial_gam]{f_binomial_gam()}}, and \code{\link[=f_binomial_rf]{f_binomial_rf()}}. This function is build for speed and it does not check the inputs.
}
\examples{
  score_auc(
    o = c(1, 1, 1, 1, 0, 0, 0),
    p = c(1, 0.8, 0.7, 0.6, 0.5, 0.6, 0.7)
  )

}
\seealso{
Other modelling_tools: 
\code{\link{case_weights}()},
\code{\link{model_formula}()},
\code{\link{score_cramer}()},
\code{\link{score_r2}()}
}
\concept{modelling_tools}
