% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.collinear_output.R
\name{short_formula}
\alias{short_formula}
\title{Shorten Formula for Concise Printing}
\usage{
short_formula(f = NULL, n = 5)
}
\arguments{
\item{f}{(required, model formula) Model formula to shorten. Default: NULL}

\item{n}{(required, integer) Number of terms to print. Default: 5}
}
\value{
character string

character string
}
\description{
Produces a compact string representation of a formula.
}
\keyword{Concise}
\keyword{Formula}
\keyword{Printing}
\keyword{Produces}
\keyword{Shorten}
\keyword{a}
\keyword{compact}
\keyword{for}
\keyword{formula.}
\keyword{internal}
\keyword{of}
\keyword{representation}
\keyword{string}
