% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_summary.collinear_output.R
\name{summary.collinear_output}
\alias{summary.collinear_output}
\title{Summary Method for Class \code{collinear_output}}
\usage{
\method{summary}{collinear_output}(x = NULL)
}
\arguments{
\item{x}{(required, list) Object of class \code{collinear_output} resulting from \code{\link[=collinear]{collinear()}}. Default: NULL}
}
\description{
Summary Method for Class \code{collinear_output}
}
