% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_summary.collinear_selection.R
\name{summary.collinear_selection}
\alias{summary.collinear_selection}
\title{Summary of \code{collinear_selection}}
\usage{
\method{summary}{collinear_selection}(object = NULL, ...)
}
\arguments{
\item{object}{(required, list of class \code{collinear_selection}) Object to summarize. Default: NULL}

\item{...}{Ignored, kept for consistency with generic.}
}
\description{
Summary of \code{collinear_selection}
}
