% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encode.R
\name{target_encode}
\alias{target_encode}
\title{Target encoding of 'df' argument}
\usage{
target_encode(df = NULL, response = NULL, predictors = NULL)
}
\arguments{
\item{df}{(required; data frame or tibble) A data frame Default: NULL.}

\item{response}{(recommended, character string) Name of a numeric response variable. Character response variables are ignored. Please, see 'Details' to better understand how providing this argument or not leads to different results when there are character variables in 'predictors'. Default: NULL.}

\item{predictors}{(optional; character vector) Character vector with column names of predictors in 'df'. If ommited, all columns of 'df' are used as predictors. Default:'NULL'}
}
\value{
data frame
}
\description{
Target encoding of 'df' argument
}
