% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encoder.R
\name{target_encoder}
\alias{target_encoder}
\title{Target encode a categorical variable based on a numeric variable.}
\usage{
target_encoder(x = NULL, y = NULL, check_input = TRUE)
}
\arguments{
\item{x}{(required; numeric vector) A numeric vector. Default: NULL}

\item{y}{(required; character vector) A character vector of the same length as 'x', representing the categorical variable.  Default: NULL}

\item{check_input}{(required; logical) If FALSE, disables data checking for a slightly better performance when many tests are performed. Default: TRUE}
}
\value{
A numeric vector containing target-encoded values.
}
\description{
This function calculates the mean of the numeric variable for each group defined by
the categorical variable and returns the target-encoded values.
}
\examples{
if(interactive()){
data(ecoregions)
encoded_values <- target_encoder(
  x = ecoregions$plant_richness,
  y = ecoregions$dominant_landcover
  )
}

}
