% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{Tolerance for Detecting Linear Dependencies}
\usage{
tolerance()
}
\value{
numeric
}
\description{
Automated definition of the \code{tol} argument in \code{\link[=solve]{solve()}}. Returns \code{0} if the system has long-double capabilities, and \code{.Machine$double.eps} otherwise.
}
\examples{
tolerance()
}
