% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_encoding_method.R
\name{validate_arg_encoding_method}
\alias{validate_arg_encoding_method}
\title{Validate Argument \code{encoding_method}}
\usage{
validate_arg_encoding_method(
  encoding_method = NULL,
  overwrite = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{encoding_method}{(optional; character or NULL). Name of one target encoding method. One of: "loo", "mean", or "rank" (see \code{\link[=target_encoding_lab]{target_encoding_lab()}} for further details). If NULL, target encoding is disabled. Default: NULL}

\item{overwrite}{(optional; logical) If TRUE, the original predictors in \code{df} are overwritten with their encoded versions, but only one encoding method, smoothing, white noise, and seed are allowed. Otherwise, encoded predictors with their descriptive names are added to \code{df}. Default: FALSE}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
character
}
\description{
Validate Argument \code{encoding_method}
}
\examples{
x <- validate_arg_encoding_method(
  encoding_method = "wrong_method"
  )
}
\seealso{
Other data_validation: 
\code{\link{autoconfig_arg_max_cor}()},
\code{\link{autoconfig_arg_max_vif}()},
\code{\link{tolerance}()},
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_f}()},
\code{\link{validate_arg_function_name}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_quiet}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
