% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_f.R
\name{validate_arg_f}
\alias{validate_arg_f}
\title{Validate Argument \code{f}}
\usage{
validate_arg_f(f = NULL, f_name = NULL, function_name = NULL)
}
\arguments{
\item{f}{(optional: function name) Unquoted name of the function to compute preference order. Available functions are listed in the column \code{name} of the dataframe returned by \code{\link[=f_functions]{f_functions()}}. By default calls to \code{\link[=f_auto]{f_auto()}} to select a suitable method depending on the nature of the data (see \code{\link[=f_auto_rules]{f_auto_rules()}}). Default: f_auto}

\item{f_name}{(optional, string) Name of the function \code{f}, as returned by \code{deparse(substitute(f))}. Default: NULL}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}
}
\value{
function
}
\description{
Validate Argument \code{f}
}
\examples{
x <- validate_arg_df(df = f_auto)
}
