% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_function_name.R
\name{validate_arg_function_name}
\alias{validate_arg_function_name}
\title{Validates Argument \code{function_name}}
\usage{
validate_arg_function_name(default_name = NULL, function_name = NULL, ...)
}
\arguments{
\item{default_name}{(optional, character) Name of the calling function. Default: NULL}

\item{function_name}{(optional, character) Name of the parent function.}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
character
}
\description{
Contatenates function names for a better message trace.
}
\seealso{
Other data_validation: 
\code{\link{autoconfig_arg_max_cor}()},
\code{\link{autoconfig_arg_max_vif}()},
\code{\link{tolerance}()},
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_encoding_method}()},
\code{\link{validate_arg_f}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_quiet}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
