% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_predictors_cor.R
\name{validate_arg_predictors_cor}
\alias{validate_arg_predictors_cor}
\title{Validation of Argument \code{predictors} for Correlation Analysis}
\usage{
validate_arg_predictors_cor(
  df = NULL,
  predictors = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. Default: NULL}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console Default: FALSE}
}
\value{
character vector: predictors names
}
\description{
Internal function to assess whether the input arguments \code{df} and \code{predictors} result in data dimensions suitable for pairwise correlation analysis. Expects argument \code{df} to be validated with \code{\link[=validate_arg_df]{validate_arg_df()}}.

The function performs the following actions:
\itemize{
\item Returns \code{predictors} immediately if it has the attribute \code{validated_cor = TRUE}.
\item If \code{predictors} does not have the attribute \code{validated = TRUE}, it is validated with \code{\link[=validate_arg_predictors]{validate_arg_predictors()}}.
\item Stops if \code{df} is \code{NULL}, using \code{\link[=validate_arg_df_not_null]{validate_arg_df_not_null()}}.
\item If \code{predictors} is empty, issues a warning and returns \code{NULL}.
\item If only one predictor is provided, returns it with a message and skips pairwise correlation filtering.
\item If fewer than 10 complete cases are available for the selected predictors, issues a warning and returns \code{NULL}.
\item Adds the attributes \code{validated = TRUE} and \code{validated_cor = TRUE} to the returned predictor vector.
}
}
\examples{
data(vi, vi_predictors)

predictors <- validate_arg_predictors_cor(
  df = vi,
  predictors = vi_predictors,
  function_name = "cor_df()"
)

  attributes(predictors)$validated
  attributes(predictors)$validated_cor
}
\seealso{
Other data_validation: 
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
