% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_predictors_vif.R
\name{validate_arg_predictors_vif}
\alias{validate_arg_predictors_vif}
\title{Validate Argument \code{predictors} for VIF Analysis}
\usage{
validate_arg_predictors_vif(
  df = NULL,
  predictors = NULL,
  preference_order = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{preference_order}{(optional; character vector, output of \code{\link[=preference_order]{preference_order()}}, or NULL). Incompatible with \code{f} (overrides it when provided). Prioritizes predictors to preserve the most relevant ones during multicollinearity filtering.

Accepted inputs are:

\itemize{

\item \strong{NULL} (default): If argument \code{f} is NULL (default), predictors are ranked from lower to higher multicollinearity. Otherwise, \code{\link[=preference_order]{preference_order()}} ranks the predictors according to their relationship with \code{responses} using the function defined in \code{f}. NOTE: The output of this setting might differ to an external call to \code{\link[=preference_order]{preference_order()}} if target encoding is triggered.

\item \strong{character vector}: Predictor names in a user-defined priority order. The first predictor in this vector is always selected, unless it has near zero-variance values. This option sets \code{f} to NULL.

\item \strong{data frame}: output of \code{\link[=preference_order]{preference_order()}} computed on the given \code{responses}. This option sets \code{f} to NULL.

\item \strong{named list}: list of data frames, output of \code{\link[=preference_order]{preference_order()}} when argument \code{responses} is a vector of length two or more. This option sets \code{f} to NULL.
}. Default: NULL}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
character vector: predictors names
}
\description{
Internal function to assess whether the input arguments \code{df} and \code{predictors} result in data dimensions suitable for a VIF analysis. Expects argument \code{df} to be validated with \code{\link[=validate_arg_df]{validate_arg_df()}}.

The function performs the following actions:
\itemize{
\item Returns \code{predictors} immediately if it has the attribute \code{validated_vif = TRUE}.
\item If \code{predictors} does not have the attribute \code{validated = TRUE}, it is validated with \code{\link[=validate_arg_predictors]{validate_arg_predictors()}}.
\item Stops if \code{df} is \code{NULL}, using \code{\link[=validate_arg_df_not_null]{validate_arg_df_not_null()}}.
\item Identifies numeric predictors using \code{\link[=identify_predictors_numeric]{identify_predictors_numeric()}}. If none are found, issues a warning and returns \code{NULL}.
\item If the number of numeric predictors is 1, it returns the input with a message and skips VIF filtering.
\item If the number of rows in \code{df} is less than 10 times the number of numeric predictors minus one, it issues a warning and returns a reduced set of predictors based on available rows. If no predictors can be retained, it returns \code{NULL}.
\item Non-numeric predictors are dropped with a message.
\item Adds the attributes \code{validated = TRUE} and \code{validated_vif = TRUE} to the returned predictor vector.
}
}
\examples{
data(vi, vi_predictors)

predictors <- validate_arg_predictors_vif(
  df = vi,
  predictors = vi_predictors,
  function_name = "vif_df()"
)

 attributes(predictors)$validated
 attributes(predictors)$validated_vif
}
\seealso{
Other data_validation: 
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
