% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_quiet.R
\name{validate_arg_quiet}
\alias{validate_arg_quiet}
\title{Validate Argument \code{quiet}}
\usage{
validate_arg_quiet(quiet = FALSE, function_name = NULL)
}
\arguments{
\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}
}
\value{
logical
}
\description{
Internal function to validate the logical argument \code{quiet}, which triggers messaging when \code{FALSE}
}
\examples{

x <- validate_arg_quiet(
  quiet = TRUE
  )

attributes(x)$validated

}
\seealso{
Other argument_validation: 
\code{\link{drop_geometry_column}()},
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_encoding_method}()},
\code{\link{validate_arg_f}()},
\code{\link{validate_arg_function_name}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_responses}()}
}
\concept{argument_validation}
