% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_response.R
\name{validate_arg_response}
\alias{validate_arg_response}
\title{Validate Argument \code{response}}
\usage{
validate_arg_response(
  df = NULL,
  response = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}. When \code{encoding_method} is not NULL, response/s are used as reference to map categorical predictors, if any, to numeric (see \code{\link[=target_encoding_lab]{target_encoding_lab()}}). When \code{f} is not NULL, responses are used to rank predictors and preserve important ones during multicollinearity filtering (see \code{\link[=preference_order]{preference_order()}}). If no response is provided, the predictors are ranked from lower to higher multicollinearity. When several responses are provided, the selection results are named after each response in the output list. If no response is provided, the variable selection shows with the name "result" in the output list. Default: NULL.}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
character string: response name
}
\description{
Internal function. Validates the argument \code{response} as follows.
\itemize{
\item Returns \code{response} as-is if it has the attribute \code{validated = TRUE}.
\item Stops if \code{df} is \code{NULL}, using \code{\link[=validate_arg_df_not_null]{validate_arg_df_not_null()}}.
\item Returns \code{NULL} immediately if \code{response} is \code{NULL}.
\item If multiple values are passed to \code{response}, uses the first and issues a warning.
\item Returns \code{NULL} with a warning if \code{response} is not a column name in \code{df}.
\item Returns \code{NULL} with a warning if \code{response} is numeric with near zero variance.
\item Returns \code{NULL} with a warning if \code{response} is non-numeric and has constant values.
\item Issues a message (if \code{quiet = FALSE}) if the response column contains \code{NA} values.
\item Adds the attribute \code{validated = TRUE} to the returned \code{response}.
}
}
\examples{

data(vi)

response <- validate_arg_response(
  df = vi,
  response = "vi_numeric"
)

attributes(response)$validated

}
\seealso{
Other data_validation: 
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
