% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_predictors}
\alias{validate_predictors}
\title{Validate Argument predictors}
\usage{
validate_predictors(
  df = NULL,
  response = NULL,
  predictors = NULL,
  decimals = 4
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{response}{(optional, character string) Name of a numeric (non-numerics are ignored) response variable in \code{df}, only required if there are categorical variables named in the argument \code{predictors}. Default: NULL.}

\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}

\item{decimals}{(required, integer) Number of decimal places for the zero variance test. Smaller numbers will increase the number of variables detected as near-zero variance. Recommended values will depend on the range of the numeric variables in 'df'. Default: 4}
}
\value{
character vector: predictor names
}
\description{
Internal function to validate the \code{predictors} argument. Requires the argument 'df' to be validated with \code{\link[=validate_df]{validate_df()}}.

Validates the 'predictors' argument to ensure it complies with the requirements of the package functions. It performs the following actions:
\itemize{
\item Stops if 'df' is NULL.
\item Stops if 'df' is not validated.
\item If 'predictors' is NULL, uses column names of 'df' as 'predictors' in the 'df' data frame.
\item Print a message if there are names in 'predictors' not in the column names of 'df', and returns only the ones in 'df'.
\item Stop if the number of numeric columns in 'predictors' is smaller than 'min_numerics'.
\item Print a message if there are zero-variance columns in 'predictors' and returns a new 'predictors' argument without them.
\item Tags the vector with the attribute \code{validated = TRUE} to let the package functions skip the data validation.
}
}
\examples{

data(
  vi,
  vi_predictors
  )

#validating example data frame
vi <- validate_df(
  df = vi
)

#validating example predictors
vi_predictors <- validate_predictors(
  df = vi,
  predictors = vi_predictors
)

#tagged as validated
attributes(vi_predictors)$validated

}
\seealso{
Other data_preparation: 
\code{\link{drop_geometry_column}()},
\code{\link{identify_non_numeric_predictors}()},
\code{\link{identify_zero_variance_predictors}()},
\code{\link{validate_df}()},
\code{\link{validate_preference_order}()},
\code{\link{validate_response}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_preparation}
