% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_preference_order}
\alias{validate_preference_order}
\title{Validate Argument preference_order}
\usage{
validate_preference_order(
  predictors = NULL,
  preference_order = NULL,
  preference_order_auto = NULL
)
}
\arguments{
\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}

\item{preference_order}{(optional; character vector) Variable names in \code{predictors}. Defines a priority order, from first to last, to preserve variables during the selection process. Variables not included in this argument are ranked by their Variance Inflation Factor. See \code{\link[=preference_order]{preference_order()}}. Default: NULL}

\item{preference_order_auto}{(required, character vector) names of the predictors in the automated preference order returned by \code{\link[=vif_select]{vif_select()}} or \code{\link[=cor_select]{cor_select()}}}
}
\value{
character vector: ranked variable names
}
\description{
Internal function to validate the argument \code{preference_order}.
}
\examples{
data(
  vi,
  vi_predictors
  )

#validating example data frame
vi <- validate_df(
  df = vi
)

#validating example predictors
vi_predictors <- validate_predictors(
  df = vi,
  predictors = vi_predictors
)

#tagged as validated
attributes(vi_predictors)$validated

#validate preference order
my_order <- c(
  "swi_max",
  "swi_min",
  "swi_deviance" #wrong one
)

my_order <- validate_preference_order(
  predictors = vi_predictors,
  preference_order = my_order,
  preference_order_auto = vi_predictors
)

#has my_order first
#excludes wrong names
#all other variables ordered according to preference_order_auto
my_order
}
\seealso{
Other data_preparation: 
\code{\link{drop_geometry_column}()},
\code{\link{identify_non_numeric_predictors}()},
\code{\link{identify_numeric_predictors}()},
\code{\link{identify_zero_variance_predictors}()},
\code{\link{validate_df}()},
\code{\link{validate_predictors}()},
\code{\link{validate_response}()}
}
\concept{data_preparation}
