% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_response}
\alias{validate_response}
\title{Validate the 'response' argument for target encoding of non-numeric variables}
\usage{
validate_response(df = NULL, response = NULL, decimals = 4)
}
\arguments{
\item{df}{(required; data frame or tibble) A validated data frame with numeric and/or character predictors predictors, and optionally, a response variable. Default: NULL.}

\item{response}{(optional, character string) Name of a numeric response variable. Character response variables are ignored. Default: NULL.}

\item{decimals}{(required, integer) number of decimal places for the zero variance test. Default: 4}
}
\value{
character string with name of the response
}
\description{
Validate the 'response' argument for target encoding of non-numeric variables
}
\examples{
if (interactive()) {
  #load example data
  #and its predictors
  data(
    vi,
    vi_predictors
    )

  #validate the vi data frame
  vi <- collinear::validate_df(df = vi)


  #validate the response "vi_mean"
  response <- collinear::validate_response(
    df = vi,
    response = "vi_mean"
    )
}
}
