% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi_predictors}
\alias{vi_predictors}
\title{All Predictor Names in Example Data Frame vi}
\format{
Character vector with predictor names.
}
\usage{
data(vi_predictors)
}
\description{
All Predictor Names in Example Data Frame vi
}
\seealso{
\link{vi}

Other example_data: 
\code{\link{toy}},
\code{\link{vi}},
\code{\link{vi_predictors_categorical}},
\code{\link{vi_predictors_numeric}}
}
\concept{example_data}
\keyword{datasets}
