% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi_predictors_numeric}
\alias{vi_predictors_numeric}
\title{Vector of numeric predictor names in \code{vi} and \code{vi_smol}}
\format{
Character vector with predictor names.
}
\usage{
data(vi_predictors_numeric)
}
\description{
Vector of numeric predictor names in \code{vi} and \code{vi_smol}
}
\seealso{
\link{vi}

Other example_data: 
\code{\link{toy}},
\code{\link{vi}},
\code{\link{vi_predictors}},
\code{\link{vi_predictors_categorical}},
\code{\link{vi_responses}},
\code{\link{vi_smol}}
}
\concept{example_data}
\keyword{datasets}
