% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi_smol}
\alias{vi_smol}
\title{Small Version (100 rows) of the Dataset \code{vi}}
\format{
Data frame with 100 rows and 68 columns.
}
\usage{
data(vi_smol)
}
\description{
Same as dataset \link{vi}, but with only 610 rows. Used in code examples to shorten runtime.
}
\seealso{
\link{vi_predictors}

Other example_data: 
\code{\link{toy}},
\code{\link{vi}},
\code{\link{vi_predictors}},
\code{\link{vi_predictors_categorical}},
\code{\link{vi_predictors_numeric}},
\code{\link{vi_responses}}
}
\concept{example_data}
\keyword{datasets}
