% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_select.R
\name{vif_select}
\alias{vif_select}
\title{Automated Multicollinearity Filtering with Variance Inflation Factors}
\usage{
vif_select(df = NULL, predictors = NULL, preference_order = NULL, max_vif = 5)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}

\item{preference_order}{(optional; character vector) Variable names in \code{predictors}. Defines a priority order, from first to last, to preserve variables during the selection process. Variables not included in this argument are ranked by their Variance Inflation Factor. See \code{\link[=preference_order]{preference_order()}}. Default: NULL}

\item{max_vif}{(optional, numeric) Maximum Variance Inflation Factor allowed during variable selection. Recommended values are between 2.5 and 10. Higher values return larger number of predictors with a higher multicollinearity. If NULL, the variance inflation analysis is disabled. Default: 5.}
}
\value{
character vector; names of selected predictors
}
\description{
This function automatizes multicollinearity filtering in data frames with numeric predictors by combining two methods:
\itemize{
\item \strong{Preference Order}: method to rank and preserve relevant variables during  multicollinearity filtering. See argument \code{preference_order} and function \code{\link[=preference_order]{preference_order()}}.
\item \strong{VIF-based filtering}: recursive algorithm to identify and remove predictors with a VIF above a given threshold.
}

When the argument \code{preference_order} is not provided, the predictors are ranked lower to higher VIF. The predictor selection resulting from this option, albeit diverse and uncorrelated, might not be the one with the highest overall predictive power when used in a model.

Please check the sections \strong{Preference Order}, \strong{Variance Inflation Factors}, and \strong{VIF-based Filtering} at the end of this help file for further details.
}
\section{Preference Order}{


The argument \code{preference_order} helps answer the question "what variable should I remove first?". Such answer is provided as a character vector of predictor names in the order of preference (from first to last) in which they want to be preserved. This vector does not need to name all predictors in \code{predictors}, but only the relevant ones for the user.

The function \code{\link[=preference_order]{preference_order()}} helps defining a quantitative criteria for preference order by assessing the relationship between each variable in \code{predictors} with a response variable.
If the argument \verb{preference order} is not provided, \code{\link[=cor_select]{cor_select()}} ranks the predictors by the sum of their absolute pairwise correlations with the other predictors, and \code{\link[=vif_select]{vif_select()}} ranks them from lower to higher VIF.
}

\section{Variance Inflation Factors}{


The Variance Inflation Factor for a given variable \eqn{a} is computed as \eqn{1/(1-R2)}, where \eqn{R2} is the multiple R-squared of a multiple regression model fitted using \eqn{a} as response and all other predictors in the input data frame as predictors, as in  \eqn{a = b + c + ...}.

The square root of the VIF of \eqn{a} is the factor by which the confidence interval of the estimate for \eqn{a} in the linear model \eqn{y = a + b + c + ...}` is widened by multicollinearity in the model predictors.

The range of VIF values is (1, Inf]. The recommended thresholds for maximum VIF may vary depending on the source consulted, being the most common values, 2.5, 5, and 10.
}

\section{VIF-based Filtering}{


The function \code{\link[=vif_select]{vif_select()}} computes Variance Inflation Factors and removes variables iteratively, until all variables in the resulting selection have a VIF below \code{max_vif}.

If the argument \code{preference_order} is not provided, all variables are ranked from lower to higher VIF, as returned by \code{\link[=vif_df]{vif_df()}}, and the variable with the higher VIF above \code{max_vif} is removed on each iteration.

If \code{preference_order} is defined, whenever two or more variables are above \code{max_vif}, the one higher in \code{preference_order} is preserved, and the next one with a higher VIF is removed. For example, for the predictors and preference order \eqn{a} and \eqn{b}, if any of their VIFs is higher than \code{max_vif}, then \eqn{b} will be removed no matter whether its VIF is lower or higher than \eqn{a}'s VIF. If their VIF scores are lower than \code{max_vif}, then both are preserved.
}

\examples{
data(vi)

#reduce data size
vi <- vi[1:1000, ]

#if argument predictors is not provided
#all numeric variables in df are used
selected.predictors <- vif_select(
  df = vi
)

selected.predictors


#otherwise, only variables in predictors are selected
selected.predictors <- vif_select(
  df = vi,
  predictors = c(
    "swi_mean",
    "rainfall_mean",
    "evapotranspiration_mean",
    "humidity_mean"
  ),
  max_vif = 2.5
)

selected.predictors

#all these have a VIF lower than max_vif (2.5)
vif_df(
  df = vi,
  predictors = selected.predictors
)


#higher max_vif results in larger selection
selected.predictors <- vif_select(
  df = vi,
  predictors = vi_predictors_numeric,
  max_vif = 10
)

selected.predictors


#smaller max_vif results in smaller selection
selected.predictors <- vif_select(
  df = vi,
  predictors = vi_predictors_numeric,
  max_vif = 2.5
)

selected.predictors


#using manual preference order
selected.predictors <- vif_select(
  df = vi,
  predictors = vi_predictors_numeric,
  preference_order = c(
    "swi_mean",
    "rainfall_mean",
    "evapotranspiration_mean",
    "humidity_mean"
  ),
  max_vif = 2.5
)

#only "humidity_mean" is lost from the results
selected.predictors


#using automated preference order
df_preference <- preference_order(
  df = vi,
  response = "vi_numeric",
  predictors = vi_predictors_numeric
)

df_preference

selected.predictors <- vif_select(
  df = vi,
  predictors = vi_predictors_numeric,
  preference_order = df_preference,
  max_vif = 10
)

selected.predictors


#categorical predictors are ignored
#the function returns NA
selected.predictors <- vif_select(
  df = vi,
  predictors = vi_predictors_categorical
)

selected.predictors


#if predictors has length 1
#selection is skipped
#and data frame with one row is returned
selected.predictors <- vif_select(
  df = vi,
  predictors = vi_predictors_numeric[1]
)

selected.predictors
}
\references{
\itemize{
\item David A. Belsley, D.A., Kuh, E., Welsch, R.E. (1980). Regression Diagnostics: Identifying Influential Data and Sources of Collinearity. John Wiley & Sons. DOI: 10.1002/0471725153.
}
}
\seealso{
Other automated_multicollinearity_analysis: 
\code{\link{collinear}()}
}
\author{
Blas M. Benito, PhD
}
\concept{automated_multicollinearity_analysis}
