% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_build.collinear_output.R
\name{build.collinear_output}
\alias{build.collinear_output}
\title{Builds Object of Class \code{collinear_output}}
\usage{
build.collinear_output(
  df = NULL,
  response = NULL,
  predictors = NULL,
  preference_order = NULL,
  selection = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}, used to target encode categorical predictors, if any, when \code{encoding_method} is not NULL, and to compute preference order when \code{preference_order} is not NULL. If \code{response} is NULL, target encoding and preference order computation are disabled. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. Default: NULL}

\item{preference_order}{(optional; character vector, output of \code{\link[=preference_order]{preference_order()}}, list, or NULL). Prioritizes predictors to preserve the most relevant ones during multicollinearity filtering. Accepted inputs are:
\itemize{
\item \strong{NULL} (default): If argument \code{f} is NULL (default), predictors are prioritized from lower to higher multicollinearity. Otherwise it calls \code{\link[=preference_order]{preference_order()}} to rank the predictors using the function defined in the argument \code{f}. The output of this setting might differ to an external call to \code{\link[=preference_order]{preference_order()}} if target encoding is triggered to transform ategorical predictors to numeric.
\item \strong{character vector}: Predictor names in a user-defined priority order. The first predictor in this vector is always selected, unless it has near zero-variance values.
\item \strong{data frame}: output of \code{\link[=preference_order]{preference_order()}} computed on the value of the argument \code{response}.
\item \strong{named list}: list of data frames, output of \code{\link[=preference_order]{preference_order()}} when argument \code{response} is a vector of length two or more.
}. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console Default: FALSE}
}
\value{
Object of class \code{collinear_output}:

\itemize{
\item If \code{response} is \code{NULL} or has length 1: an object of class \code{collinear_output} with a custom print method \link{print.collinear_output}. It contains:
\itemize{
\item \code{response} (\code{character} or \code{NULL}): Name of the response variable.
\item \code{predictors} (\code{character} or \code{NULL}): Names of predictors considered in multicollinearity filtering.
\item \code{selection} (\code{character} or \code{NULL}): Names of selected, non-collinear predictors.
\item \code{df} (\code{data.frame} or \code{NULL}): Data frame including columns in \code{response} and \code{selection}. If target encoding was applied, categorical variables will be numeric.
\item \code{arguments} (\code{list}): A list of arguments used in the call: \code{encoding_method}, \code{preference_order}, \code{f}, \code{max_cor}, and \code{max_vif}
}

\item If \code{response} is a character vector of length >= 2: an object of class \code{collinear_list}, a named list of \code{"collinear_output"} objects, one per response variable.
}
}
\description{
Builds Object of Class \code{collinear_output}
}
