% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_weights.R
\name{case_weights}
\alias{case_weights}
\title{Case Weights for Unbalanced Binomial or Categorical Responses}
\usage{
case_weights(x = NULL)
}
\arguments{
\item{x}{(required, integer, character, or factor vector) binomial, categorical, or factor response variable. Default: \code{NULL}}
}
\value{
numeric vector: case weights
}
\description{
Case Weights for Unbalanced Binomial or Categorical Responses
}
\examples{
 case_weights(
   x = c(0, 0, 0, 1, 1)
   )

 case_weights(
   x = c("a", "a", "b", "b", "c")
   )
}
\seealso{
Other preference_order_tools: 
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{f_functions}()},
\code{\link{preference_order_collinear}()},
\code{\link{preference_score_auc}()},
\code{\link{preference_score_r2}()},
\code{\link{preference_score_v}()}
}
\concept{preference_order_tools}
