% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{case_weights}
\alias{case_weights}
\title{Case Weights for Unbalanced Binomial or Categorical Responses}
\usage{
case_weights(x = NULL)
}
\arguments{
\item{x}{(required, integer, character, or factor vector) Either a binomial response with 1s and 0s, or categorical responses as characters or factors. Default: \code{NULL}}
}
\value{
numeric vector: case weights
}
\description{
Case Weights for Unbalanced Binomial or Categorical Responses
}
\examples{
if(interactive()){

 case_weights(
   x = c(0, 0, 0, 1, 1)
 )

 case_weights(
   x = c("a", "a", "b", "b", "c")
 )

 }
}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{f_auc}},
\code{\link{f_cramerv}()},
\code{\link{f_cramerv_rf_categorical}()},
\code{\link{f_r2}},
\code{\link{f_r2_counts}},
\code{\link{preference_order}()}
}
\concept{preference_order}
