% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_class.collinear_selection.R
\name{class.collinear_selection}
\alias{class.collinear_selection}
\title{Builds \code{class.collinear_selection}}
\usage{
class.collinear_selection(
  df = NULL,
  response = NULL,
  preference_order = NULL,
  selection = NULL,
  quiet = FALSE,
  function_name = NULL,
  ...
)
}
\arguments{
\item{df}{(required; dataframe, tibble, or sf) A dataframe with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional, character string) Name of a response variable in \code{df}. Default: NULL.}

\item{preference_order}{(optional; character vector, dataframe resulting from \code{\link[=preference_order]{preference_order()}}, or NULL). Prioritizes predictors to preserve the most relevant ones during multicollinearity filtering. Overrides \code{f} when not NULL.}

\item{selection}{(required, character vector) Names of the selected predictors. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{...}{(optional) Used to pass internal arguments such as \code{function_name} for \link{validate_arg_function_name} in nested calls, or \code{m}, a correlation matrix generated via \code{\link[stats:cor]{stats::cor()}} or \code{\link[=cor_matrix]{cor_matrix()}}.}
}
\value{
list
\itemize{
\item \code{df} (dataframe):  input data with the given \code{response}, if any, and the selected \code{predictors}. If target encoding was triggered, then categorical predictors will show as numeric in this dataframe.
\item \code{response} (character): name of the response, if any.
\item \code{selection} (character vector): names of the selected \code{predictors}.
\item \code{formulas} (list): if \code{responses} is not NULL, list with modelling formulas generated by \code{\link[=model_formula]{model_formula()}} on the given response and the selected predictors. When the response and the predictors are numeric, the formulas "linear" and "smooth" (for GAM models) are generated. When the response is categorical, a "classification" formula is returned.
\item{preference_order} (list): list with two elements, the dataframe \code{df}, generated by \code{\link[=preference_order]{preference_order()}}, and the list \code{f} wiht the name, the expression, and the metric used by the function \code{f} used to compute preference order.
}
}
\description{
Builds \code{class.collinear_selection}
}
\seealso{
Other S3_methods: 
\code{\link{print.collinear_output}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_output}()},
\code{\link{summary.collinear_selection}()}
}
\concept{S3_methods}
