% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{collinear-package}
\alias{collinear-package}
\title{collinear}
\description{
Package for multicollinearity management in data frames with numeric and categorical variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://blasbenito.github.io/collinear/}
  \item Report bugs at \url{https://github.com/blasbenito/collinear/issues}
}

}
\author{
\strong{Maintainer}: Blas M. Benito \email{blasbenito@gmail.com} (\href{https://orcid.org/0000-0001-5105-7232}{ORCID}) [copyright holder]

}
\keyword{internal}
