% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinear_auto.R
\name{collinear_auto}
\alias{collinear_auto}
\title{Automated Multicollinearity Management}
\usage{
collinear_auto(df = NULL, responses = NULL, predictors = NULL, quiet = FALSE)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{responses}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}. When \code{encoding_method} is not NULL, response/s are used as reference to map categorical predictors, if any, to numeric (see \code{\link[=target_encoding_lab]{target_encoding_lab()}}). When \code{f} is not NULL, responses are used to rank predictors and preserve important ones during multicollinearity filtering (see \code{\link[=preference_order]{preference_order()}}). If no response is provided, the predictors are ranked from lower to higher multicollinearity. When several responses are provided, the selection results are named after each response in the output list. If no response is provided, the variable selection shows with the name "result" in the output list. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
list of class \code{class.collinear_output}
}
\description{
A fire-and-forget wrapper to \code{\link[=collinear]{collinear()}}. It configures target encoding, preference order, and the maximum multicollinearity levels (\code{max_vif} and \code{max_cor}) automatically.

Target encoding is triggered with the method "loo" (leave-one-out, see \code{\link[=target_encoding_lab]{target_encoding_lab()}} for further details) when a response is numeric and there are categorical predictors.

Preference order is computed automatically when \code{responses} is not NULL using \code{\link[=f_auto]{f_auto()}} to select a proper method depending on the nature of the response and the predictors.

Arguments \code{max_vif} and \code{max_cor} are autoconfigured depending on the average correlation between all pairs of predictors. The autoconfiguration rules are shown below:

\preformatted{

max_cor <- max(
  0.5,
  correlation_mean
)

if (max_cor >= 0.95) {
  max_vif <- 10
} else if (max_cor >= 0.85) {
  max_vif <- 8.5
} else if (max_cor >= 0.80) {
  max_vif <- 7
} else if (max_cor >= 0.75) {
  max_vif <- 5
} else if (max_cor >= 0.70) {
  max_vif <- 3.5
} else {
  max_vif <- 2.5
}

}

These rules were obtained empirically, by comparing executions of \code{\link[=vif_select]{vif_select()}} and \code{\link[=cor_select]{cor_select()}} on 100000 different combinations of rows and columns of the dataset \link{vi}, and finding the \code{max_vif} value leading to the most similar variable selection resulting from a given \code{max_cor} value.
}
\examples{

data(
  vi_smol,
  vi_predictors_numeric
)

x <- collinear_auto(
  df = vi_smol,
  responses = "vi_numeric",
  predictors = vi_predictors_numeric
)

x

}
\seealso{
Other automated_multicollinearity_analysis: 
\code{\link{collinear}()},
\code{\link{collinear_lite}()}
}
\concept{automated_multicollinearity_analysis}
