% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinear_auto.R
\name{collinear_auto}
\alias{collinear_auto}
\title{Automated Multicollinearity Management}
\usage{
collinear_auto(
  df = NULL,
  responses = NULL,
  predictors = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{(required; dataframe, tibble, or sf) A dataframe with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{responses}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df}. If NULL, all columns in \code{df} except the ones in \code{responses} and those with constant values or near zero variance are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Used to pass internal arguments such as \code{function_name} for \link{validate_arg_function_name} in nested calls, or \code{m}, a correlation matrix generated via \code{\link[stats:cor]{stats::cor()}} or \code{\link[=cor_matrix]{cor_matrix()}}.}
}
\value{
list of class \code{collinear_output} with sub-lists of class \code{collinear_selection}.

If \code{responses = NULL}, only one sub-list named "result" is produced. Otherwise, a sub-list named after each response is generated. A sub-list of class \code{collinear_selection} contains:
\itemize{
\item \code{df} (dataframe):  input data with the given \code{response}, if any, and the selected \code{predictors}. If target encoding was triggered, then categorical predictors will show as numeric in this dataframe.
\item \code{response} (character): name of the response, if any.
\item \code{selection} (character vector): names of the selected \code{predictors}.
\item \code{formulas} (list): if \code{responses} is not NULL, list with modelling formulas generated by \code{\link[=model_formula]{model_formula()}} on the given response and the selected predictors. When the response and the predictors are numeric, the formulas "linear" and "smooth" (for GAM models) are generated. When the response is categorical, a "classification" formula is returned.
\item{preference_order} (list): list with two elements, the dataframe \code{df}, generated by \code{\link[=preference_order]{preference_order()}}, and the list \code{f} with the name, the expression, and the metric used by the function \code{f} used to compute preference order.
}
}
\description{
A fire-and-forget wrapper to \code{\link[=collinear]{collinear()}} to automatically manage multicollinearity without user input.

It configures target encoding, preference order, and the maximum multicollinearity levels (\code{max_vif} and \code{max_cor}) automatically.

Target encoding is triggered with the method "loo" (leave-one-out, see \code{\link[=target_encoding_lab]{target_encoding_lab()}} for further details) when a response is numeric and there are categorical predictors.

Preference order is computed automatically when \code{responses} is not NULL. The function \code{\link[=f_auto]{f_auto()}} is used to select a proper method depending on the response and predictor types.

Configures \code{max_cor} as \code{max(x, 0.58)}, where \code{x} is the median correlation between all predictors as returned by \code{\link[=cor_stats]{cor_stats()}}. The minimum correlation \code{0.58} roughly matches a VIF of 2.5.

Configures \code{max_vif} as the prediction of the model \link{gam_cor_to_vif} on the given \code{max_cor} value obtained as explained above.

In the experiments performed during development (see \link{experiment_collinear_auto}), the function reduced the total number of predictors by 68\%, with a range between 80\% for 100 initial predictors and 44\% for 25 predictors.

The average difference in median correlation between the input and selected predictors was -0.16, and the resulting maximum VIF across all selections was 2.75, with an average of 2.3.

These results indicate that the function offers a robust functionality with a minimal setup.
}
\examples{

data(
  vi_smol,
  vi_predictors_numeric
)

x <- collinear_auto(
  df = vi_smol,
  responses = "vi_numeric",
  predictors = vi_predictors_numeric
)

x

}
\seealso{
Other automated_multicollinearity_analysis: 
\code{\link{collinear}()},
\code{\link{collinear_select}()},
\code{\link{collinear_stats}()}
}
\concept{automated_multicollinearity_analysis}
