% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_clusters.R
\name{cor_clusters}
\alias{cor_clusters}
\title{Hierarchical Clustering from a Pairwise Correlation Matrix}
\usage{
cor_clusters(
  df = NULL,
  predictors = NULL,
  max_cor = 0.75,
  method = "complete",
  plot = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.
TODO: rename to responses, and check what other functions inherit this param}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{max_cor}{(optional; numeric or NULL) Maximum correlation allowed between pairs of \code{predictors}. Valid values are between 0.01 and 0.99, and recommended values are between 0.5 (strict) and 0.9 (permissive). If NULL, the pairwise correlation analysis is disabled. Default: 0.75}

\item{method}{(optional, character string) Argument of \code{\link[stats:hclust]{stats::hclust()}} defining the agglomerative method. One of: "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). Unambiguous abbreviations are accepted as well. Default: "complete".}

\item{plot}{(optional, logical) If TRUE, the clustering is plotted. Default: FALSE}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
data frame: predictor names and their clusters
}
\description{
Hierarchical clustering of predictors from their pairwise correlation matrix. Computes the correlation matrix with \code{\link[=cor_df]{cor_df()}} and \code{\link[=cor_matrix]{cor_matrix()}}, transforms it to a dist object, computes a clustering solution with \code{\link[stats:hclust]{stats::hclust()}}, and applies \code{\link[stats:cutree]{stats::cutree()}} to separate groups based on the value of the argument \code{max_cor}.

Returns a data frame with predictor names and their clusters, and optionally, prints a dendrogram of the clustering solution.

Accepts a parallelization setup via \code{\link[future:plan]{future::plan()}} and a progress bar via \code{\link[progressr:handlers]{progressr::handlers()}} (see examples).
}
\examples{
  data(vi)

  #subset to speed-up example
  vi <- vi[1:1000, ]

  #OPTIONAL: parallelization setup
  # future::plan(
  #   future::multisession,
  #   workers = 2
  # )

  #OPTIONAL: progress bar
  # progressr::handlers(global = TRUE)

  #group predictors using max_cor as clustering threshold
  df_clusters <- cor_clusters(
    df = vi,
    predictors = c(
      "koppen_zone", #character
      "soil_type", #factor
      "topo_elevation", #numeric
      "soil_temperature_mean" #numeric
    ),
    max_cor = 0.75,
    plot = FALSE #set to TRUE to plot result
  )

  #OPTIONAL: disable parallelization
  #future::plan(future::sequential)
}
\seealso{
Other pairwise_correlation: 
\code{\link{cor_cramer_v}()},
\code{\link{cor_df}()},
\code{\link{cor_matrix}()},
\code{\link{cor_select}()}
}
\concept{pairwise_correlation}
