% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_df.R
\name{cor_df}
\alias{cor_df}
\alias{cor_numeric_vs_numeric}
\alias{cor_numeric_vs_categorical}
\alias{cor_categorical_vs_categorical}
\title{Absolute Pairwise Correlations Dataframe}
\usage{
cor_df(df = NULL, predictors = NULL, quiet = FALSE, ...)

cor_numeric_vs_numeric(df = NULL, predictors = NULL, quiet = FALSE, ...)

cor_numeric_vs_categorical(df = NULL, predictors = NULL, quiet = FALSE, ...)

cor_categorical_vs_categorical(
  df = NULL,
  predictors = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
data frame:
\itemize{
\item \code{x} predictor name.
\item \code{y} predictor name.
\item \code{correlation} Absolute Pearson correlation for numeric vs. numeric and numeric vs. categorical, or Cramer's V for categorical vs categorical.
}
}
\description{
Computes a absolute pairwise correlation data frame. Implements methods to compare different types of predictors:
\itemize{
\item \strong{numeric vs. numeric}: as computed with \code{\link[stats:cor]{stats::cor()}} using the methods "pearson" or "spearman", via \code{\link[=cor_numeric_vs_numeric]{cor_numeric_vs_numeric()}}.
\item \strong{numeric vs. categorical}: the function \code{\link[=cor_numeric_vs_categorical]{cor_numeric_vs_categorical()}} target-encodes the categorical variable using the numeric variable as reference with \code{\link[=target_encoding_lab]{target_encoding_lab()}} and the method "loo" (leave-one-out), and then their correlation is computed with \code{\link[stats:cor]{stats::cor()}}.
\item \strong{categorical vs. categorical}: the function \code{\link[=cor_categorical_vs_categorical]{cor_categorical_vs_categorical()}} computes Cramer's V (see \code{\link[=cor_cramer_v]{cor_cramer_v()}}) as indicator of the association between character or factor variables. Please read the documentation of \code{\link[=cor_cramer_v]{cor_cramer_v()}} to better understand the caveats of mixing Pearson Correlation and Cramer's V in a multicollinearity analysis.
}

Accepts a parallelization setup via \code{\link[future:plan]{future::plan()}} and a progress bar via \code{\link[progressr:handlers]{progressr::handlers()}} (see examples).
}
\examples{
data(vi_smol)

#predictors
predictors = c(
  "koppen_zone", #character
  "soil_type", #factor
  "topo_elevation", #numeric
  "soil_temperature_mean" #numeric
)

#OPTIONAL: parallelization setup
# only worth it for large data
# future::plan(
#   future::multisession,
#   workers = 2
# )
#
#OPTIONAL: progress bar
# progressr::handlers(global = TRUE)

#predictors ordered from lower to higher multicollinearity
x <- cor_df(
  df = vi_smol,
  predictors = predictors
)

x

#OPTIONAL: disable parallelization
#future::plan(future::sequential)
}
\seealso{
Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer_v}()},
\code{\link{cor_matrix}()},
\code{\link{cor_select}()},
\code{\link{cor_stats}()}

Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer_v}()},
\code{\link{cor_matrix}()},
\code{\link{cor_select}()},
\code{\link{cor_stats}()}

Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer_v}()},
\code{\link{cor_matrix}()},
\code{\link{cor_select}()},
\code{\link{cor_stats}()}

Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer_v}()},
\code{\link{cor_matrix}()},
\code{\link{cor_select}()},
\code{\link{cor_stats}()}
}
\concept{pairwise_correlation}
