% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_df.R
\name{cor_df}
\alias{cor_df}
\title{Pairwise correlation}
\usage{
cor_df(df = NULL, predictors = NULL, method = NULL)
}
\arguments{
\item{df}{(required; data frame or tibble) A data frame with predictors. Default: \code{NULL}.}

\item{predictors}{(optional; character vector) Character vector with the names of the predictive variables. Every element of this vector must be in the column names of \code{df}. Default: \code{NULL}}

\item{method}{(optional; charater string) Method used to compute pairwise correlations. Accepted methods are "pearson" (with a recommended minimum of 30 rows in 'df') or "spearman" (with a recommended minimum of 10 rows in 'df'). Default: "pearson".}
}
\value{
Data frame with pairs of predictors and their correlation.
}
\description{
Returns the pairwise Pearson correlation between all pairs of predictors in a training dataset.

Please note that near-zero variance columns are ignored by this function. Use \code{\link[=mc_auto_vif]{mc_auto_vif()}} to remove them.
}
\examples{
if(interactive()){

data(
  ecoregions,
  ecoregions_predictors
)

df <- cor_df(
      df = ecoregions,
      predictors = ecoregions_predictors
)

}
}
