% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_df.R
\name{cor_df}
\alias{cor_df}
\alias{cor_numerics}
\alias{cor_numerics_and_characters}
\alias{cor_characters}
\title{Pairwise Correlation Data Frame}
\usage{
cor_df(
  df = NULL,
  response = NULL,
  predictors = NULL,
  cor_method = "pearson",
  encoding_method = "mean"
)

cor_numerics(df = NULL, predictors = NULL, cor_method = "pearson")

cor_numerics_and_characters(
  df = NULL,
  predictors = NULL,
  cor_method = "pearson",
  encoding_method = "mean"
)

cor_characters(df, predictors)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{response}{(optional, character string) Name of a numeric (non-numerics are ignored) response variable in \code{df}, only required if there are categorical variables named in the argument \code{predictors}. Default: NULL.}

\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}

\item{cor_method}{(optional; character string) Method used to compute pairwise correlations. Either "pearson" or "spearman". Default: "pearson".}

\item{encoding_method}{(optional; character string). Name of the target encoding method to transform non-numeric variables to numeric. One of "mean", "rank", "loo", "rnorm" (see \code{\link[=target_encoding_lab]{target_encoding_lab()}} for further details). Default: "mean"}
}
\value{
data frame; variable pairs and their correlation

data frame

data frame

data frame
}
\description{
Returns a correlation data frame between all pairs of predictors in a training data frame. If the argument  Non-numeric predictors are transformed into numeric via target encoding, using the 'response' variable as reference.

Attempts to handle correlations between pairs of variables of different types, as follows:
\itemize{
\item numeric vs. numeric: computed with stats::cor() with the methods "pearson" or "spearman" using \code{\link[=cor_numerics]{cor_numerics()}}.
\item numeric vs. character via \code{\link[=cor_numerics_and_characters]{cor_numerics_and_characters()}}, two alternatives leading to different results:
\itemize{
\item 'response' is provided: the character variable is target-encoded as numeric using the values of the response as reference, and then its correlation with the numeric variable is computed with stats::cor(). This option generates a response-specific result suitable for training statistical and machine-learning models
\item 'response' is NULL (or the name of a non-numeric column): the character variable is target-encoded as numeric using the values of the numeric predictor (instead of the response) as reference, and then their correlation is computed with stats::cor(). This option leads to a response-agnostic result suitable for clustering problems.
}
\item character vs. character, via \code{\link[=cor_characters]{cor_characters()}}, with two alternatives leading to different results:
\itemize{
\item 'response' is provided: the character variables are target-encoded as numeric using the values of the response as reference, and then their correlation is computed with stats::cor().
\item response' is NULL (or the name of a non-numeric column): the association between the character variables is computed using Cramer's V. This option might be problematic, because R-squared values and Cramer's V, even when having the same range between 0 and 1, are not fully comparable.
}
}
}
\examples{

data(
  vi,
  vi_predictors
)

#reduce size of vi to speed-up example execution
vi <- vi[1:1000, ]
vi_predictors <- vi_predictors[1:10]

#without response
#categorical vs categorical compared with cramer_v()
#categorical vs numerical compared wit stats::cor() via target-encoding
#numerical vs numerical compared with stats::cor()
df <- cor_df(
  df = vi,
  predictors = vi_predictors
)

head(df)

#with response
#different solution than previous one
#because target encoding is done against the response
#rather than against the other numeric in the pair
df <- cor_df(
  df = vi,
  response = "vi_mean",
  predictors = vi_predictors
)

head(df)

}
\seealso{
Other correlation: 
\code{\link{cor_matrix}()},
\code{\link{cor_select}()},
\code{\link{cramer_v}()}
}
\author{
Blas M. Benito, PhD
}
\concept{correlation}
\keyword{internal}
