% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation matrix}
\usage{
cor_matrix(df = NULL, method = NULL)
}
\arguments{
\item{df}{(required; data frame or tibble) A data frame with predictors. Default: \code{NULL}.}

\item{method}{(optional; charater string) Method used to compute pairwise correlations. Accepted methods are "pearson" (with a recommended minimum of 30 rows in 'df') or "spearman" (with a recommended minimum of 10 rows in 'df'). Default: "pearson".}
}
\value{
Data frame with pairs of predictors and their correlation.
}
\description{
Returns a correlation matrix between all pairs of predictors in a training dataset. Non-numeric predictors are transformed into numeric via target encoding.
}
\examples{
if(interactive()){

data(
  ecoregions,
  ecoregions_predictors
)

df <- cor_df(
      df = ecoregions,
      predictors = ecoregions_predictors
)

}
}
