% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Pairwise Correlation Matrix}
\usage{
cor_matrix(df = NULL, predictors = NULL, quiet = FALSE, ...)
}
\arguments{
\item{df}{(required; dataframe, tibble, or sf) A dataframe with predictors or the output of \code{\link[=cor_df]{cor_df()}}. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df}. If NULL, all columns in \code{df} except the ones in \code{responses} and those with constant values or near zero variance are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Used to pass internal arguments such as \code{function_name} for \link{validate_arg_function_name} in nested calls, or \code{m}, a correlation matrix generated via \code{\link[stats:cor]{stats::cor()}} or \code{\link[=cor_matrix]{cor_matrix()}}, and the cross-validation arguments of \code{\link[=preference_order]{preference_order()}}: \code{cv_iterations}, and \code{cv_training_fraction}.}
}
\value{
correlation matrix
}
\description{
Computes a square matrix of pairwise correlations for a set of numeric and/or categorical predictors.

If \code{df} is already a correlation dataframe generated by \code{\link[=cor_df]{cor_df()}}), the function transforms it into a correlation matrix. Otherwise, \code{\link[=cor_df]{cor_df()}} is used internally to compute pairwise correlations before generating the matrix.

Supports parallel computation via \code{\link[future:plan]{future::plan()}} and optional progress reporting via \code{\link[progressr:handlers]{progressr::handlers()}}.
}
\examples{
data(vi_smol)

## OPTIONAL: parallelization setup
## irrelevant when all predictors are numeric
## only worth it for large data with many categoricals
# future::plan(
#   future::multisession,
#   workers = future::availableCores() - 1
# )

## OPTIONAL: progress bar
# progressr::handlers(global = TRUE)

predictors <- c(
  "koppen_zone", #character
  "soil_type", #factor
  "topo_elevation", #numeric
  "soil_temperature_mean" #numeric
)

#from dataframe with predictors
x <- cor_matrix(
  df = vi_smol,
  predictors = predictors
)

x

#from correlation dataframe
x <- cor_df(
  df = vi,
  predictors = predictors
) |>
  cor_matrix()

x

## OPTIONAL: disable parallelization
#future::plan(future::sequential)
}
\seealso{
Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer}()},
\code{\link{cor_df}()},
\code{\link{cor_select}()},
\code{\link{cor_stats}()}
}
\author{
Blas M. Benito, PhD
}
\concept{pairwise_correlation}
