% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Pairwise Correlation Matrix}
\usage{
cor_matrix(df = NULL, predictors = NULL, quiet = FALSE)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.
TODO: rename to responses, and check what other functions inherit this param}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
correlation matrix
}
\description{
If argument 'df' results from \code{\link[=cor_df]{cor_df()}}, transforms it to a correlation matrix. If argument 'df' is a dataframe with predictors, and the argument 'predictors' is provided then \code{\link[=cor_df]{cor_df()}} is used to compute pairwise correlations, and the result is transformed to matrix.

Accepts a parallelization setup via \code{\link[future:plan]{future::plan()}} and a progress bar via \code{\link[progressr:handlers]{progressr::handlers()}} (see examples).
}
\examples{
    data(vi)

    #subset to speed-up example
    vi <- vi[1:1000, ]

    predictors <- c(
      "koppen_zone", #character
      "soil_type", #factor
      "topo_elevation", #numeric
      "soil_temperature_mean" #numeric
    )

    #OPTIONAL: parallelization setup
    # future::plan(
    #   future::multisession,
    #   workers = 2
    # )

    #OPTIONAL: progress bar
    # progressr::handlers(global = TRUE)


    #correlation data frame to matrix
    m <- cor_df(
      df = vi,
      predictors = predictors
    ) |>
      cor_matrix()

    m

    #direct computation (uses cor_df() internally)
    m <- cor_matrix(
      df = vi,
      predictors = predictors
    )

    m

    #OPTIONAL: disable parallelization
    #future::plan(future::sequential)
}
\seealso{
Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer_v}()},
\code{\link{cor_df}()},
\code{\link{cor_select}()}
}
\author{
Blas M. Benito, PhD
}
\concept{pairwise_correlation}
