% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Absolute Pairwise Correlation Matrix}
\usage{
cor_matrix(df = NULL, predictors = NULL, quiet = FALSE, ...)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A dataframe with predictors or the output of\code{\link[=cor_df]{cor_df()}}. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
correlation matrix
}
\description{
Computes a square matrix of absolute pairwise correlations for a set of numeric and/or categorical predictors.

If \code{df} is already a correlation data frame (as returned by \code{\link[=cor_df]{cor_df()}}), the function converts it directly into a correlation matrix. Otherwise, \code{\link[=cor_df]{cor_df()}} is used internally to compute absolute pairwise correlations before creating the matrix.

Supports parallel computation via \code{\link[future:plan]{future::plan()}} and optional progress reporting via \code{\link[progressr:handlers]{progressr::handlers()}}.
}
\examples{
data(vi_smol)

predictors <- c(
  "koppen_zone", #character
  "soil_type", #factor
  "topo_elevation", #numeric
  "soil_temperature_mean" #numeric
)

#OPTIONAL: parallelization setup
# future::plan(
#   future::multisession,
#   workers = 2
# )

#OPTIONAL: progress bar
# progressr::handlers(global = TRUE)

#from dataframe with predictors
m <- cor_matrix(
  df = vi_smol,
  predictors = predictors
)

m

#from correlation dataframe
m <- cor_df(
  df = vi,
  predictors = predictors
) |>
  cor_matrix()

m

#OPTIONAL: disable parallelization
#future::plan(future::sequential)
}
\seealso{
Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer_v}()},
\code{\link{cor_df}()},
\code{\link{cor_select}()},
\code{\link{cor_stats}()}
}
\author{
Blas M. Benito, PhD
}
\concept{pairwise_correlation}
