% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_select.R
\name{cor_select}
\alias{cor_select}
\title{Automated multicollinearity reduction via pairwise correlation}
\usage{
cor_select(
  df = NULL,
  response = NULL,
  predictors = NULL,
  preference_order = NULL,
  method = "pearson",
  max_cor = 0.75
)
}
\arguments{
\item{df}{(required; data frame or tibble) A data frame with numeric and/or character predictors predictors, and optionally, a response variable. Default: NULL.}

\item{response}{(recommended, character string) Name of a numeric response variable. Character response variables are ignored. Please, see 'Details' to better understand how providing this argument or not leads to different results when there are character variables in 'predictors'. Default: NULL.}

\item{predictors}{(optional; character vector) Character vector with predictor names in 'df'. If omitted, all columns of 'df' are used as predictors. Default:'NULL'}

\item{preference_order}{(optional; character vector) vector with column names in 'predictors' in the desired preference order, or result of the function \code{preference_order()}. Allows defining a priority order for selecting predictors, which can be particularly useful when some predictors are more critical for the analysis than others. Predictors not included in this argument are ranked by their Variance Inflation Factor. Default: \code{NULL}.}

\item{method}{(optional; character string) Method used to compute pairwise correlations. Accepted methods are "pearson" (with a recommended minimum of 30 rows in 'df') or "spearman" (with a recommended minimum of 10 rows in 'df'). Default: "pearson".}

\item{max_cor}{(optional; numeric) Maximum correlation allowed between any pair of predictors. Higher values return larger number of predictors with higher multicollinearity. Default: \code{0.75}}
}
\value{
Character vector with the names of the selected predictors.
}
\description{
Applies a recursive algorithm to remove variables with a bivariate correlation with any other variable higher than a threshold defined by the argument \code{max_cor}.

If the argument \code{response} is provided, all non-numeric variables in \code{predictors} are transformed into numeric using target encoding (see \code{target_encode()}). Otherwise, non-numeric variables are ignored.

The argument \code{preference_order} allows defining a preference selection order to preserve (when possible) variables that might be interesting or even required for a given analysis

For example, if \code{predictors} is \verb{"c("a", "b", "c")} and \code{preference_order} is \verb{"c("a", "b")}, there are two possibilities:
\itemize{
\item If the correlation between \code{"a"} and \code{"b"} is below \code{max_cor}, both variables are selected.
\item If their correlation is equal or above \code{max_cor}, then \code{"a"} is selected, no matter its correlation with \code{"c"},
}

If \code{preference_order} is not provided, then the predictors are ranked by their variance inflation factor as computed by \code{vif_df()}.
}
\examples{
if(interactive()){

 #loading example data
 data(
  vi,
  vi_predictors
  )

 #without preference order
 selected.variables <- cor_select(
   df = vi,
   predictors = vi_predictors
 )

 selected.variables

 #with preference order
 selected.variables <- cor_select(
   df = vi,
   predictors = vi_predictors,
   preference_order = vi_predictors[1:5],
 )

 selected.variables

}
}
