% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_stats.R
\name{cor_stats}
\alias{cor_stats}
\title{Pairwise Correlation Stats}
\usage{
cor_stats(df = NULL, predictors = NULL, quiet = FALSE)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows. Alternatively, a dataframe resulting from \code{\link[=cor_df]{cor_df()}}. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
list:
\itemize{
\item values: numeric vector of absolute correlations.
\item stats: dataframe with the columns \code{statistic} and \code{value}.
}
}
\description{
Generates a correlation dataframe with \code{\link[=cor_df]{cor_df()}} and computes the the minimum, mean, maximum, and quantiles 0.05, 0.25, 0.50, 0.75, and 0.95 of the Pearson correlations (and Cramer's V if there is more than one categorical predictor).
}
\examples{
stats <- cor_stats(
  df = vi_smol,
  predictors = vi_predictors_numeric
)

head(stats$values)
str(stats$stats)
}
\seealso{
Other pairwise_correlation: 
\code{\link{cor_clusters}()},
\code{\link{cor_cramer_v}()},
\code{\link{cor_df}()},
\code{\link{cor_matrix}()},
\code{\link{cor_select}()}
}
\concept{pairwise_correlation}
