% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_geometry_column.R
\name{drop_geometry_column}
\alias{drop_geometry_column}
\title{Removes geometry column in sf data frames}
\usage{
drop_geometry_column(df)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}
}
\value{
The input data frame without a geometry column
}
\description{
Replicates the functionality of  \verb{\link[sf]\{st_drop_geometry\}} without depending on the \code{sf} package.
}
\seealso{
Other data_preparation: 
\code{\link{identify_non_numeric_predictors}()},
\code{\link{identify_zero_variance_predictors}()},
\code{\link{validate_df}()},
\code{\link{validate_predictors}()},
\code{\link{validate_preference_order}()},
\code{\link{validate_response}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_preparation}
