% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_geometry_column.R
\name{drop_geometry_column}
\alias{drop_geometry_column}
\title{Removes geometry column in sf data frames}
\usage{
drop_geometry_column(df = NULL, quiet = FALSE, ...)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
data frame
}
\description{
Replicates the functionality of \code{sf::st_drop_geometry()} without depending on the \code{sf} package.
}
\examples{
data(vi_smol)

#creating fake geometry column without sf loaded
vi_smol$geometry <- NA
attr(
  x = vi_smol,
  which = "sf_column"
  ) <- "geometry"

#check new attribute
attributes(vi_smol)$sf_column

#drop geometry column
df <- drop_geometry_column(
  df = vi_smol
  )

#checking that the geometry was droppped
"geometry" \%in\% colnames(df)
attributes(df)$sf_column

}
\author{
Blas M. Benito, PhD
}
\concept{data_preparation}
