% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_geometry_column.R
\name{drop_geometry_column}
\alias{drop_geometry_column}
\title{Removes geometry column in sf data frames}
\usage{
drop_geometry_column(df = NULL, quiet = FALSE)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{quiet}{(optional; logical) If FALSE, messages generated during the execution of the function are printed to the console Default: FALSE}
}
\value{
data frame
}
\description{
Replicates the functionality of  \code{sf::st_drop_geometry()} without depending on the \code{sf} package.
}
\seealso{
Other data_preparation: 
\code{\link{identify_predictors}()},
\code{\link{identify_predictors_categorical}()},
\code{\link{identify_predictors_numeric}()},
\code{\link{identify_predictors_zero_variance}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_preparation}
