% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_geometry_column.R
\name{drop_geometry_column}
\alias{drop_geometry_column}
\title{Removes geometry column in sf data frames}
\usage{
drop_geometry_column(df)
}
\arguments{
\item{df}{(required; data frame or matrix) Input data frame. Default: NULL}
}
\value{
The input data frame without a geometry column
}
\description{
Replicates the functionality of  \verb{\link[sf]\{st_drop_geometry\}} without depending on the \code{sf} package.
}
\author{
Blas M. Benito
}
\keyword{internal}
