% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_auto}
\alias{f_auto}
\title{Automatic Selection of Function to Compute Preference Order}
\usage{
f_auto(df = NULL, response = NULL, predictors = NULL, quiet = FALSE, ...)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional, character string) Name of a response variable in \code{df}. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
function name
}
\description{
Internal function to select a proper \code{f_...()} function to compute preference order depending on the types of the response variable and the predictors. The selection criteria is available as a data frame generated by \code{\link[=f_auto_rules]{f_auto_rules()}}.
}
\examples{
data(vi, vi_predictors_numeric)
f <- f_auto(
  df = vi[1:1000, ],
  response = "vi_numeric",
  predictors = vi_predictors_numeric
  )
}
\seealso{
Other preference_order_tools: 
\code{\link{f_auto_rules}()},
\code{\link{f_functions}()}
}
\concept{preference_order_tools}
