% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_auto_rules.R
\name{f_auto_rules}
\alias{f_auto_rules}
\title{Rules Used by \code{f_auto()} to Select an \code{f} Function}
\usage{
f_auto_rules()
}
\value{
dataframe
}
\description{
Dataframe with rules used by \code{\link[=f_auto]{f_auto()}} to select the function \code{f} in \code{\link[=f_functions]{f_functions()}} to compute preference order in \code{\link[=preference_order]{preference_order()}}. In most cases, random forest is selected as base model to provide homogeneous results across case types.
}
\examples{
f_auto_rules()
}
\seealso{
Other preference_order_tools: 
\code{\link{f_auto}()},
\code{\link{f_functions}()}
}
\concept{preference_order_tools}
