% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_auto_rules}
\alias{f_auto_rules}
\title{Rules to Select Default f Argument to Compute Preference Order}
\usage{
f_auto_rules()
}
\value{
data frame
}
\description{
Data frame with rules used by \code{\link[=f_auto]{f_auto()}} to select the function \code{f} to compute preference order in \code{\link[=preference_order]{preference_order()}}.
}
\examples{
f_auto_rules()
}
\seealso{
Other preference_order_tools: 
\code{\link{f_auto}()},
\code{\link{f_functions}()},
\code{\link{preference_order_collinear}()}
}
\concept{preference_order_tools}
