% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_functions.R
\name{f_functions}
\alias{f_functions}
\title{List predictor scoring functions}
\usage{
f_functions()
}
\value{
dataframe
}
\description{
List predictor scoring functions
}
\examples{
f_functions()
}
\seealso{
Other preference_order_tools: 
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()}
}
\concept{preference_order_tools}
