% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_methods.R
\name{f_functions}
\alias{f_functions}
\title{Data Frame of Preference Functions}
\usage{
f_functions()
}
\value{
data frame
}
\description{
Data Frame of Preference Functions
}
\examples{
f_functions()
}
\seealso{
Other preference_order_tools: 
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{preference_order_collinear}()}
}
\concept{preference_order_tools}
