% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_gam_deviance}
\alias{f_gam_deviance}
\title{Explained Deviance From a GAM Model}
\usage{
f_gam_deviance(x, y, df)
}
\arguments{
\item{x}{(required, character string) name of the predictor variable.}

\item{y}{(required, character string) name of the response variable}

\item{df}{(required, data frame) data frame with the columns 'x' and 'y'.}
}
\value{
numeric: explained deviance
}
\description{
Computes the explained deviance of a response against a predictor via Generalized Additive Model (GAM). This option is slower than \code{\link[=f_rsquared]{f_rsquared()}}, but suitable when fitting GAMs with the results of a multicollinearity filtering.
}
\examples{

data(vi)

#subset to limit example run time
vi <- vi[1:1000, ]

#this example requires "mgcv" installed in the system
if(
  requireNamespace(
    package = "mgcv",
   quietly = TRUE
   )
 ){

  f_gam_deviance(
    x = "growing_season_length", #predictor
    y = "vi_mean",               #response
    df = vi
  )

}

}
\seealso{
Other preference_order: 
\code{\link{auc_score}()},
\code{\link{case_weights}()},
\code{\link{f_gam_auc_balanced}()},
\code{\link{f_gam_auc_unbalanced}()},
\code{\link{f_logistic_auc_balanced}()},
\code{\link{f_logistic_auc_unbalanced}()},
\code{\link{f_rf_auc_balanced}()},
\code{\link{f_rf_auc_unbalanced}()},
\code{\link{f_rf_rsquared}()},
\code{\link{f_rsquared}()},
\code{\link{preference_order}()}
}
\concept{preference_order}
