% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_rf_deviance}
\alias{f_rf_deviance}
\title{Compute the preference order for predictors based on the random forest R-squared value.}
\usage{
f_rf_deviance(x, y, df)
}
\arguments{
\item{x}{The predictor variable name.}

\item{y}{The response variable name.}

\item{df}{The data frame containing the variables.}
}
\value{
The R-squared value from the Random Forest model for the predictor.
}
\description{
This function calculates the preference order of predictors based on the R-squared value from a Random Forest model between each predictor and the response variable.
}
