% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_rf_deviance}
\alias{f_rf_deviance}
\title{R-squared of Random Forest model from out-of-bag data}
\usage{
f_rf_deviance(x, y, df)
}
\arguments{
\item{x}{(required, character string) name of the predictor variable.}

\item{y}{(required, character string) name of the response variable}

\item{df}{(required, data frame) data frame with the columns 'x' and 'y'.}
}
\value{
R-squared
}
\description{
Computes a univariate random forest model with \verb{\link[ranger]\{ranger\}} and returns the R-squared on the out-of-bag data.
}
\examples{
if(interactive()){

data(vi)

f_rf_deviance(
  x = "growing_season_length", #predictor
  y = "vi_mean",               #response
  df = vi
)

}
}
