% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{f_rsquared}
\alias{f_rsquared}
\title{Compute the preference order for predictors based on the R-squared value.}
\usage{
f_rsquared(x, y, df)
}
\arguments{
\item{x}{The predictor variable name.}

\item{y}{The response variable name.}

\item{df}{The data frame containing the variables.}
}
\value{
The R-squared value between the predictor and the response variable.
}
\description{
This function calculates the preference order of predictors based on the R-squared value between each predictor and the response variable.
}
