% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gam_cor_to_vif}
\alias{gam_cor_to_vif}
\title{GAM Relating Maximum Correlation to VIF Threshold}
\format{
A \code{\link[mgcv]{gam}} object.
}
\source{
Generated internally from \link{experiment_cor_vs_vif}.
}
\usage{
gam_cor_to_vif
}
\description{
A fitted generalized additive model describing \code{max_vif} as a function of \code{max_cor} in \link{experiment_cor_vs_vif}.
}
\details{
The model was fitted with a smooth term in \code{max_cor} (basis dimension \code{k = 9}) and cubic weights proportional to the Jaccard similarity between the results of \code{\link[=cor_select]{cor_select()}} and \code{\link[=vif_select]{vif_select()}} to emphasize cases with high agreement between the two selection methods.

\preformatted{
gam_cor_to_vif <- mgcv::gam(
 formula = output_max_vif ~ s(input_max_cor, k = 9),
 weights = experiment_cor_vs_vif$out_selection_jaccard^3,
 data = experiment_cor_vs_vif,
 select = TRUE
 )
}
Model performance:
\itemize{
\item Adjusted R-squared: 0.918
\item Deviance explained: 91.8\\%
\item Estimated scale: 0.238
\item Effective degrees of freedom for smooth: 7.47 (Ref.df = 8)
\item F-statistic for smooth term: 13,101 (p < 2e-16)
}
}
\examples{
data(gam_cor_to_vif)
plot(gam_cor_to_vif, shade = TRUE)
}
\keyword{datasets}
