% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_non_numeric_predictors}
\alias{identify_non_numeric_predictors}
\title{Identify Non-Numeric Predictors}
\usage{
identify_non_numeric_predictors(df = NULL, predictors = NULL)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}
}
\value{
character vector: names of non-numeric predictors
}
\description{
Returns the names of the non-numeric predictors, if any, in a data frame.
}
\examples{

data(
  vi,
  vi_predictors
)

non.numeric.predictors <- identify_non_numeric_predictors(
  df = vi,
  predictors = vi_predictors
)

non.numeric.predictors

}
\seealso{
Other data_preparation: 
\code{\link{drop_geometry_column}()},
\code{\link{identify_numeric_predictors}()},
\code{\link{identify_zero_variance_predictors}()},
\code{\link{validate_df}()},
\code{\link{validate_predictors}()},
\code{\link{validate_preference_order}()},
\code{\link{validate_response}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_preparation}
