% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_predictors.R
\name{identify_predictors}
\alias{identify_predictors}
\title{Identify Numeric, Categorical, and Logical Predictors}
\usage{
identify_predictors(
  df = NULL,
  predictors = NULL,
  decimals = 4,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{decimals}{(required, integer) Number of decimal places for the zero variance test. Smaller numbers will increase the number of variables detected as near-zero variance. Recommended values will depend on the range of the numeric variables in 'df'. Default: 4}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
list
\itemize{
\item \code{numeric}: character vector of numeric predictors.
\item \code{categorical}: character vector of categorical (character and factor) predictors.
\item \code{logical}: character vector of logical predictors.
\item \code{zero_variance}: character vector with numeric predictors having quasi-constant values
}
}
\description{
Returns a list with the names of the valid numeric, categorical, and logical predictors in a given dataframe
}
\examples{

data(vi, vi_predictors)

#adding logical predictor
vi$logical <- TRUE

#adding zero variance predictor
vi$zero_variance <- 1

#identify predictors types
x <- identify_predictors(
  df = vi
)

x

#identify specific predictors
x <- identify_predictors(
  df = vi,
  predictors = c("logical", "zero_variance")
)

x

}
\seealso{
Other data_types: 
\code{\link{identify_predictors_categorical}()},
\code{\link{identify_predictors_logical}()},
\code{\link{identify_predictors_numeric}()},
\code{\link{identify_predictors_type}()},
\code{\link{identify_predictors_zero_variance}()},
\code{\link{identify_response_type}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_types}
