% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_response_type}
\alias{identify_response_type}
\title{Identify Response Type}
\usage{
identify_response_type(df = NULL, response = NULL, quiet = FALSE)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.
TODO: rename to responses, and check what other functions inherit this param}

\item{response}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}. When \code{encoding_method} is not NULL, response/s are used as reference to map categorical predictors, if any, to numeric (see \code{\link[=target_encoding_lab]{target_encoding_lab()}}). When \code{f} is not NULL, responses are used to rank predictors and preserve important ones during multicollinearity filtering (see \code{\link[=preference_order]{preference_order()}}). If no response is provided, the predictors are ranked from lower to higher multicollinearity. When several responses are provided, the selection results are named after each response in the output list. If no response is provided, the variable selection shows with the name "result" in the output list. Default: NULL.}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
character string: response type
}
\description{
Internal function to identify the type of response variable. Supported types are:
\itemize{
\item "continuous-binary": decimal numbers and two unique values; results in a warning, as this type is difficult to model.
\item "continuous-low": decimal numbers and 3 to 5 unique values; results in a message, as this type is difficult to model.
\item "continuous-high": decimal numbers and more than 5 unique values.
\item "integer-binomial": integer with 0s and 1s, suitable for binomial models.
\item "integer-binary": integer with 2 unique values other than 0 and 1; returns a warning, as this type is difficult to model.
\item "integer-low": integer with 3 to 5 unique values or meets specified thresholds.
\item "integer-high": integer with more than 5 unique values suitable for count modelling.
\item "categorical": character or factor with 2 or more levels.
\item "unknown": when the response type cannot be determined.
}
}
\examples{
data(vi, vi_predictors)

identify_response_type(
  df = vi,
  response = "vi_numeric"
)

identify_response_type(
  df = vi,
  response = "vi_counts"
)

identify_response_type(
  df = vi,
  response = "vi_binomial"
)

identify_response_type(
  df = vi,
  response = "vi_categorical"
)

identify_response_type(
  df = vi,
  response = "vi_factor"
)

}
\seealso{
Other data_types: 
\code{\link{identify_predictors}()},
\code{\link{identify_predictors_categorical}()},
\code{\link{identify_predictors_logical}()},
\code{\link{identify_predictors_numeric}()},
\code{\link{identify_predictors_type}()},
\code{\link{identify_predictors_zero_variance}()}
}
\concept{data_types}
