% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_data}
\alias{check_data}
\alias{check_predictors_names}
\alias{check_response_name}
\alias{non_numeric_columns}
\alias{zero_variance_columns}
\alias{numeric_columns}
\title{Checks 'data' argument}
\usage{
check_data(data = NULL, drop.geometry = FALSE, verbose = TRUE)

check_predictors_names(
  predictors.names = NULL,
  data = NULL,
  is.required = TRUE,
  numeric.only = TRUE,
  na.allowed = FALSE,
  zero.variance.allowed = FALSE,
  decimal.places = 4,
  verbose = TRUE
)

check_response_name(
  response.name = NULL,
  data = NULL,
  is.required = TRUE,
  na.allowed = FALSE,
  zero.variance.allowed = FALSE,
  decimal.places = 4,
  verbose = TRUE
)

non_numeric_columns(data, columns = NULL)

zero_variance_columns(data, columns = NULL, decimal.places = 4)

numeric_columns(data = NULL, columns = NULL)
}
\arguments{
\item{data}{A data frame to extract numeric columns from.}

\item{verbose}{logical}

\item{predictors.names}{predictors.names.argument}

\item{is.required}{logical}

\item{numeric.only}{logical}

\item{na.allowed}{logical, changes the check depending on whether NAs are allowed in data or not.}

\item{zero.variance.allowed}{logical}

\item{decimal.places}{Integer, number of decimal places to round \code{columns} to. Defines the tolerance of the test. Default: 4}

\item{response.name}{response.name}

\item{columns}{A character vector specifying the columns of the data frame to extract.}

\item{drop.gemetry}{drops geometry column if data is an sf data frame}
}
\value{
data

predictor.names

response.name

A character vector with the names of the numeric columns in the selected data frame.

A character vector with the names of the numeric columns in the selected data frame.

A character vector with the names of the non-numeric columns in the selected data frame.
}
\description{
This function takes a data frame and a set of columns and returns the names of the numeric columns in the selected data frame.

This function takes a data frame and a set of columns and returns the names of the columns with zero variance.

This function takes a data frame and a set of columns and returns the names of the non-numeric columns in the selected data frame.
}
\keyword{internal}
