% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{case_weights}
\alias{case_weights}
\title{Case Weights for Unbalanced Binary Response}
\usage{
case_weights(x = NULL)
}
\arguments{
\item{x}{(required, integer vector) Binary response with 1s and 0s. Default: \code{NULL}}
}
\value{
A vector with a length equal to \code{x} with case weights.
}
\description{
Case Weights for Unbalanced Binary Response
}
\examples{
if(interactive()){

 case_weights(
   x = c(0, 0, 0, 1, 1)
 )

 }
}
\keyword{internal}
