% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_score_v.R
\name{performance_score_v}
\alias{performance_score_v}
\title{Cramer's V of Observations vs Predictions}
\usage{
performance_score_v(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required; character vector) character vector representing observations of a categorical variable. Default: NULL}

\item{p}{(required; character vector) character vector representing predictions of a categorical variable. Must have the same length as \code{o}. Default: NULL}
}
\value{
numeric: Cramer's V
}
\description{
Internal function to compute the Cramer's V of categorical observations versus categorical model predictions.
}
\examples{
performance_score_v(
 o = c("a", "a", "b", "c", "c"),
 p = c("a", "b", "b", "c", "c")
 )

}
\seealso{
Other modelling_tools: 
\code{\link{case_weights}()},
\code{\link{model_formula}()},
\code{\link{performance_score_auc}()},
\code{\link{performance_score_r2}()}
}
\concept{modelling_tools}
