% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_score_v.R
\name{performance_score_v}
\alias{performance_score_v}
\title{Cramer's V of Observations vs Predictions}
\usage{
performance_score_v(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required, numeric vector) Response values. Default: NULL}

\item{p}{(required, numeric vector) Model predictions. Default: NULL}
}
\value{
numeric: Cramer's V
}
\description{
Internal function to compute the Cramer's V of categorical observations versus categorical model predictions.
}
\seealso{
Other modelling_tools: 
\code{\link{case_weights}()},
\code{\link{model_formula}()},
\code{\link{performance_score_auc}()},
\code{\link{performance_score_r2}()}
}
\concept{modelling_tools}
