% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{preference_order}
\alias{preference_order}
\title{Compute the preference order for predictors based on a user-defined function.}
\usage{
preference_order(
  df = NULL,
  response = NULL,
  predictors = NULL,
  f = f_rsquared,
  workers = 1
)
}
\arguments{
\item{df}{(required; data frame or tibble) A data frame with numeric and/or character predictors predictors, and optionally, a response variable. Default: NULL.}

\item{response}{(recommended, character string) Name of a numeric response variable. Character response variables are ignored. Please, see 'Details' to better understand how providing this argument or not leads to different results when there are character variables in 'predictors'. Default: NULL.}

\item{predictors}{(optional; character vector) character vector with predictor names in 'df'. If omitted, all columns of 'df' are used as predictors. Default:'NULL'}

\item{f}{(optional: function) A function that returns a value representing the relationship between a given predictor and the response. Higher values are ranked higher. The available options are:
\itemize{
\item f_rsquared (default option): returns the R-squared of the correlation between the response and the predictor.
\item f_lm_coef: returns the absolute coefficient of a linear model of the response against the scaled predictor.
\item f_gam_deviance: returns the explained deviance of a GAM model of the response against a predictor. Requires the package \code{mgcv}.
\item f_f_variance: returns the explained deviance of a random forest model of the response against the predictor. Requires the package \code{ranger}.
}}

\item{workers}{(integer) number of workers for parallel execution. Default: 1}
}
\value{
A data frame with the columns "predictor" and "value". The former contains the predictors names in order, ready for the argument \code{preference_order} in \code{cor_select()}, \code{vif_select()} and \code{collinear()}. The latter contains the result of the function \code{f} for each combination of predictor and response.
}
\description{
This function calculates the preference order of predictors based on a user-provided function that takes a predictor, a response, and a data frame as arguments.
}
\examples{
if(interactive()){

data(
  vi,
  vi_predictors
)

preference.order <- preference_order(
  df = vi,
  response = "vi_mean",
  predictors = vi_predictors[1:10],
  f = f_rsquared,
  workers = 2
  )

#without preference_order
selected.predictors <- collinear(
  df = vi,
  predictors = vi_predictors[1:10],
  preference_order = preference.order$predictor,
  max_cor = 0.75,
  max_vif = 5
  )

selected.predictors

}

}
