% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{preference_order}
\alias{preference_order}
\title{Quantitative Variable Prioritization for Multicollinearity Filtering}
\usage{
preference_order(df = NULL, response = NULL, predictors = NULL, f = NULL)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{response}{(optional, character string) Name of a numeric (non-numerics are ignored) response variable in \code{df}, only required if there are categorical variables named in the argument \code{predictors}. Default: NULL.}

\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}

\item{f}{(optional: function) Function to compute preference order. If NULL (default), the output of \code{\link[=f_default]{f_default()}} for the given data is used:
\itemize{
\item \code{\link[=f_auc_rf]{f_auc_rf()}}: \code{response} is binomial.
\item \code{\link[=f_r2_pearson]{f_r2_pearson()}}: \code{response} and \code{predictors} are numeric.
\item \code{\link[=f_v]{f_v()}}: \code{response} and \code{predictors} are categorical.
\item \code{\link[=f_v_rf_categorical]{f_v_rf_categorical()}}: \code{response} is categorical and \code{predictors} are numeric or mixed .
\item \code{\link[=f_r2_rf]{f_r2_rf()}}: in all other cases.
}
Default: NULL}
}
\value{
data frame
}
\description{
Ranks a set of predictors by the strength of their association with a response. Aims to minimize the loss of important predictors during multicollinearity filtering.

The strength of association between the response and each predictor is computed by the function \code{f}. The \code{f} functions available are:
\itemize{
\item \strong{Numeric response vs numeric predictor}:
\itemize{
\item \code{\link[=f_r2_pearson]{f_r2_pearson()}}: Pearson's R-squared.
\item \code{\link[=f_r2_spearman]{f_r2_spearman()}}: Spearman's R-squared.
\item \code{\link[=f_r2_glm_gaussian]{f_r2_glm_gaussian()}}: Pearson's R-squared of response versus the predictions of a Gaussian GLM.
\item \code{\link[=f_r2_glm_gaussian_poly2]{f_r2_glm_gaussian_poly2()}}: Gaussian GLM with second degree polynomial.
\item \code{\link[=f_r2_gam_gaussian]{f_r2_gam_gaussian()}}: GAM model fitted with \code{\link[mgcv:gam]{mgcv::gam()}}.
\item \code{\link[=f_r2_rpart]{f_r2_rpart()}}: Recursive Partition Tree fitted with \code{\link[rpart:rpart]{rpart::rpart()}}.
\item \code{\link[=f_r2_rf]{f_r2_rf()}}: Random Forest model fitted with \code{\link[ranger:ranger]{ranger::ranger()}}.
}
\item \strong{Integer counts response vs. numeric predictor}:
\itemize{
\item \code{\link[=f_r2_glm_poisson]{f_r2_glm_poisson()}}: Pearson's R-squared of a Poisson GLM.
\item \code{\link[=f_r2_glm_poisson_poly2]{f_r2_glm_poisson_poly2()}}: Poisson GLM with second degree polynomial.
\item \code{\link[=f_r2_gam_poisson]{f_r2_gam_poisson()}}: Poisson GAM.
}
\item \strong{Binomial response (1 and 0) vs. numeric predictor}:
\itemize{
\item \code{\link[=f_auc_glm_binomial]{f_auc_glm_binomial()}}: AUC of quasibinomial GLM with weighted cases.
\item \code{\link[=f_auc_glm_binomial_poly2]{f_auc_glm_binomial_poly2()}}: As above with second degree polynomial.
\item \code{\link[=f_auc_gam_binomial]{f_auc_gam_binomial()}}: Quasibinomial GAM with weighted cases.
\item \code{\link[=f_auc_rpart]{f_auc_rpart()}}: Recursive Partition Tree with weighted cases.
\item \code{\link[=f_auc_rf]{f_auc_rf()}}: Random Forest model with weighted cases.
}
\item \strong{Categorical response (character of factor) vs. categorical predictor}:
\itemize{
\item \code{\link[=f_v]{f_v()}}: Cramer's V between two categorical variables.
}
\item \strong{Categorical response vs. categorical or numerical predictor}:
\itemize{
\item \code{\link[=f_v_rf_categorical]{f_v_rf_categorical()}}: Cramer's V of a Random Forest model.
}
}

The name of the used function is stored in the attribute "f_name" of the output data frame. It can be retrieved via \code{attributes(df)$f_name}

Additionally, any custom function accepting a data frame with the columns "x" (predictor) and "y" (response) and returning a numeric indicator of association where higher numbers indicate higher association will work.

This function returns a data frame with the column "predictor", with predictor names ordered by the column "preference", with the result of \code{f}. This data frame, or the column "predictor" alone, can be used as inputs for the argument \code{preference_order} in \code{\link[=collinear]{collinear()}}, \code{\link[=cor_select]{cor_select()}}, and \code{\link[=vif_select]{vif_select()}}.

This function supports a parallelization setup via \code{\link[future:plan]{future::plan()}} and progress bars generated by the \code{progressr} package. See examples.
}
\examples{
data(
  vi,
  vi_predictors,
  vi_predictors_numeric
)

#subsets to limit example run time
vi <- vi[1:1000, ]
vi_predictors <- vi_predictors[1:10]
vi_predictors_numeric <- vi_predictors_numeric[1:10]

#parallelization setup
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
# progressr::handlers(global = TRUE)

#numeric response and predictors
#applies f_r2_pearson() to compute correlation between response and predictors
df_preference <- preference_order(
  df = vi,
  response = "vi_numeric",
  predictors = vi_predictors_numeric
  )

#returns data frame ordered by preference
df_preference

#name of preference function stored as attribute
attributes(df_preference)$f_name

#using it as input in collinear()
selection <- collinear(
  df = vi,
  response = "vi_numeric",
  predictors = vi_predictors_numeric,
  preference_order = df_preference
)

selection

#numeric response and mixed predictors
#uses f_r2_rf() to compute preference
preference_order(
  df = vi,
  response = "vi_numeric",
  predictors = vi_predictors
)

#user defined f function
#for binomial response and numeric predictors
preference_order(
  df = vi,
  response = "vi_binomial",
  predictors = vi_predictors_numeric,
  f = f_auc_glm_binomial
)

#default f function
#for binomial response and numeric predictors
#uses f_auc_rf()
preference_order(
  df = vi,
  response = "vi_binomial",
  predictors = vi_predictors_numeric,
  f = NULL
)

#disable parallelization
future::plan(future::sequential)
}
\seealso{
Other preference_order: 
\code{\link{f_auc}},
\code{\link{f_r2}},
\code{\link{f_r2_counts}},
\code{\link{f_v}()},
\code{\link{f_v_rf_categorical}()}
}
\author{
Blas M. Benito, PhD
}
\concept{preference_order}
