% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order.R
\name{preference_order}
\alias{preference_order}
\title{Quantitative Variable Prioritization Multicollinearity Filtering}
\usage{
preference_order(
  df = NULL,
  response = NULL,
  predictors = NULL,
  f = f_r2_pearson,
  encoding_method = "mean"
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{response}{(optional, character string) Name of a numeric (non-numerics are ignored) response variable in \code{df}, only required if there are categorical variables named in the argument \code{predictors}. Default: NULL.}

\item{predictors}{(optional; character vector) Names of the variables to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. If argument \code{response} is not provided, non-numeric variables are ignored. Default: NULL}

\item{f}{(optional: function) Function to compute the association between each predictor in \code{predictors} and \code{response}. Default: \code{f_r2_pearson}}

\item{encoding_method}{(optional; character string). Name of the target encoding method to transform non-numeric variables to numeric. One of "mean", "rank", "loo", "rnorm" (see \code{\link[=target_encoding_lab]{target_encoding_lab()}} for further details). Default: "mean"}
}
\value{
data frame
}
\description{
This function ranks a set of variables by their association with a response to minimize loss of predictive power during multicollinearity filtering.

Returns a data frame with the columns "predictor", with predictor names in order of preference, and "preference", the result of a function \code{f} computing the association between each predictor and the response. This data frame, or the column "predictor" can be used as inputs for the argument \code{preference_order} in \code{\link[=cor_select]{cor_select()}}, \code{\link[=vif_select]{vif_select()}} and \code{\link[=collinear]{collinear()}}.

The \code{f} functions available are:
\itemize{
\item Numeric response vs numeric predictor:
\itemize{
\item \code{\link[=f_r2_pearson]{f_r2_pearson()}}: Pearson's R-squared.
\item \code{\link[=f_r2_spearman]{f_r2_spearman()}}: Spearman's R-squared.
\item \code{\link[=f_r2_glm_gaussian]{f_r2_glm_gaussian()}}: Pearson's R-squared of response versus the predictions of a Gaussian GLM.
\item \code{\link[=f_r2_glm_gaussian_poly2]{f_r2_glm_gaussian_poly2()}}: Gaussian GLM with second degree polynomial.
\item \code{\link[=f_r2_gam_gaussian]{f_r2_gam_gaussian()}}: GAM model fitted with \code{\link[mgcv:gam]{mgcv::gam()}}.
\item \code{\link[=f_r2_rpart]{f_r2_rpart()}}: Recursive Partition Tree fitted with \code{\link[rpart:rpart]{rpart::rpart()}}.
\item \code{\link[=f_r2_rf]{f_r2_rf()}}: Random Forest model fitted with \code{\link[ranger:ranger]{ranger::ranger()}}.
}
\item Integer counts response vs. numeric predictor:
\itemize{
\item \code{\link[=f_r2_glm_poisson]{f_r2_glm_poisson()}}: Pearson's R-squared of a Poisson GLM.
\item \code{\link[=f_r2_glm_poisson_poly2]{f_r2_glm_poisson_poly2()}}: Poisson GLM with second degree polynomial.
\item \code{\link[=f_r2_gam_poisson]{f_r2_gam_poisson()}}: Poisson GAM.
}
\item Binomial response (1 and 0) vs. numeric predictor:
\itemize{
\item \code{\link[=f_auc_glm_binomial]{f_auc_glm_binomial()}}: AUC of quasibinomial GLM with weighted cases.
\item \code{\link[=f_auc_glm_binomial_poly2]{f_auc_glm_binomial_poly2()}}: As above with second degree polynomial.
\item \code{\link[=f_auc_gam_binomial]{f_auc_gam_binomial()}}: Quasibinomial GAM with weighted cases.
\item \code{\link[=f_auc_rpart]{f_auc_rpart()}}: Recursive Partition Tree with weighted cases.
\item \code{\link[=f_auc_rf]{f_auc_rf()}}: Random Forest model with weighted cases.
}
\item Categorical response (character of factor) vs. categorical predictor:
\itemize{
\item \code{\link[=f_v]{f_v()}}: Cramer's V between two categorical variables.
}
\item Categorical response vs. categorical or numerical predictor:
\itemize{
\item \code{\link[=f_v_rf_categorical]{f_v_rf_categorical()}}: Cramer's V of a Random Forest model.
}
}

When the argument \code{response} is the name of a numeric column in \code{df}, then \code{\link[=target_encoding_lab]{target_encoding_lab()}} is used to transform any categorical variables named in \code{predictors} to numeric.

This function supports a parallelization setup via \code{\link[future:plan]{future::plan()}} and progress bars generated by the \code{progressr} package. See examples.
}
\examples{


data(
  vi,
  vi_predictors,
  vi_predictors_category
)

#subsets to limit example run time
vi <- vi[1:1000, ]
vi_predictors <- vi_predictors[1:10]
vi_predictors_category <- vi_predictors_category[1:10]

#parallelization setup
future::plan(
  future::multisession,
  workers = 2 #set to parallelly::availableWorkers() - 1
)

#progress bar
# progressr::handlers(global = TRUE)

#numeric response
df_preference <- preference_order(
  df = vi,
  response = "vi_numeric",
  predictors = vi_predictors,
  f = f_r2_pearson
  )

df_preference

#using it as input in collinear()
selection <- collinear(
  df = vi,
  response = "vi_numeric",
  predictors = vi_predictors,
  preference_order = df_preference
)

selection


#count response
df_preference <- preference_order(
  df = vi,
  response = "vi_counts",
  predictors = vi_predictors,
  f = f_r2_glm_poisson
)

df_preference

#binomial response
df_preference <- preference_order(
  df = vi,
  response = "vi_binomial",
  predictors = vi_predictors,
  f = f_auc_glm_binomial
)

df_preference

#categorical response and predictors
df_preference <- preference_order(
  df = vi,
  response = "vi_category",
  predictors = vi_predictors_category,
  f = f_v
)

df_preference

#categorical response and categorical and numeric predictors
df_preference <- preference_order(
  df = vi,
  response = "vi_category",
  predictors = vi_predictors,
  f = f_v_rf_categorical
)

df_preference


#disable parallelization
future::plan(future::sequential)

}
\seealso{
Other preference_order: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{default_f}()},
\code{\link{default_f_heuristic}()},
\code{\link{f_auc}},
\code{\link{f_r2}},
\code{\link{f_r2_counts}},
\code{\link{f_v}()},
\code{\link{f_v_rf_categorical}()},
\code{\link{r2}()}
}
\author{
Blas M. Benito, PhD
}
\concept{preference_order}
