% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_wrapper.R
\name{preference_order_wrapper}
\alias{preference_order_wrapper}
\title{Internal Wrapper of \code{preference_order()} Within \code{collinear()}}
\usage{
preference_order_wrapper(
  df = NULL,
  response = NULL,
  predictors = NULL,
  preference_order = NULL,
  f = NULL,
  f_name = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}, used to target encode categorical predictors, if any, when \code{encoding_method} is not NULL, and to compute preference order when \code{preference_order} is not NULL. If \code{response} is NULL, target encoding and preference order computation are disabled. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. Default: NULL}

\item{preference_order}{(optional; character vector, output of \code{\link[=preference_order]{preference_order()}}, list, or NULL). Prioritizes predictors to preserve the most relevant ones during multicollinearity filtering. Accepted inputs are:
\itemize{
\item \strong{NULL} (default): If argument \code{f} is NULL (default), predictors are prioritized from lower to higher multicollinearity. Otherwise it calls \code{\link[=preference_order]{preference_order()}} to rank the predictors using the function defined in the argument \code{f}. The output of this setting might differ to an external call to \code{\link[=preference_order]{preference_order()}} if target encoding is triggered to transform ategorical predictors to numeric.
\item \strong{character vector}: Predictor names in a user-defined priority order. The first predictor in this vector is always selected, unless it has near zero-variance values.
\item \strong{data frame}: output of \code{\link[=preference_order]{preference_order()}} computed on the value of the argument \code{response}.
\item \strong{named list}: list of data frames, output of \code{\link[=preference_order]{preference_order()}} when argument \code{response} is a vector of length two or more.
}. Default: NULL}

\item{f}{(optional: function name or NULL) Function to compute preference order used when \code{preference_order = "auto"}. Available functions are listed by \code{\link[=f_functions]{f_functions()}} and described in the manual of \code{\link[=preference_order]{preference_order()}}. If NULL, calls to \code{\link[=f_auto]{f_auto()}} (see \code{\link[=f_auto_rules]{f_auto_rules()}}) to select a suitable method depending on the nature of the data. Default: NULL}

\item{f_name}{(optional, string) Name of the function \code{f}, as returned by \code{deparse(substitute(f))}. Default: NULL}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console Default: FALSE}
}
\value{
character vector or NULL
}
\description{
Internal function to manage the argument \code{preference_order} within \code{\link[=collinear]{collinear()}}. Users should use \code{\link[=preference_order]{preference_order()}} instead.
}
\seealso{
Other preference_order_tools: 
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{f_functions}()}
}
\concept{preference_order_tools}
