% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.collinear_arguments.R
\name{print.collinear_arguments}
\alias{print.collinear_arguments}
\title{Print \code{collinear_arguments}}
\usage{
\method{print}{collinear_arguments}(x = NULL, n = 5, ...)
}
\arguments{
\item{x}{(required, list of class \code{collinear_arguments}) Object to print. Default: NULL}

\item{n}{(optional, integer) Maximum printed vector length. Default: 5.}

\item{...}{Ignored, kept for consistency with generic.}
}
\description{
Print \code{collinear_arguments}
}
\seealso{
Other S3_methods: 
\code{\link{build.collinear_arguments}()},
\code{\link{build.collinear_output}()},
\code{\link{build.collinear_selection}()},
\code{\link{print.collinear_output}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_output}()},
\code{\link{summary.collinear_selection}()}
}
\concept{S3_methods}
