% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.collinear_arguments.R
\name{print.collinear_arguments}
\alias{print.collinear_arguments}
\title{Print Method for Class \code{collinear_arguments}}
\usage{
\method{print}{collinear_arguments}(x = NULL, n = 5)
}
\arguments{
\item{x}{(required, list) Object of class \code{collinear_arguments}. Default: NULL}

\item{n}{(optional, integer) Maximum number of vector elements to print. Default: 5.}
}
\description{
Prints an object of the class \code{collinear_arguments} produced by \code{\link[=collinear]{collinear()}}.
}
