% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.collinear_output.R
\name{print.collinear_output}
\alias{print.collinear_output}
\title{Print \code{class.collinear_output}}
\usage{
\method{print}{collinear_output}(x = NULL, n = 5, ...)
}
\arguments{
\item{x}{(required, list of class \code{class.collinear_output}) Object to print. Default: NULL}

\item{n}{(optional, integer) Maximum printed vector length. Default: 5.}

\item{...}{Ignored, kept for consistency with generic.}
}
\description{
Print \code{class.collinear_output}
}
\seealso{
Other S3_methods: 
\code{\link{class.collinear_selection}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_output}()},
\code{\link{summary.collinear_selection}()}
}
\concept{S3_methods}
