% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_print.collinear_selection.R
\name{print.collinear_selection}
\alias{print.collinear_selection}
\title{Print Method for Class \code{collinear_selection}}
\usage{
\method{print}{collinear_selection}(x = NULL, n = 5)
}
\arguments{
\item{x}{(required, list) Object of class \code{collinear_selection} resulting from \code{\link[=build.collinear_selection]{build.collinear_selection()}}. Default: NULL}

\item{n}{(optional, integer) Maximum number of vector elements to print. Defaults: 5.}
}
\description{
Prints an obect of the class \link{collinear_selection} produced by \code{\link[=build.collinear_selection]{build.collinear_selection()}}.
}
