% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_order_tools.R
\name{r2}
\alias{r2}
\title{Pearson's R-squared of Observations vs Predictions}
\usage{
r2(o = NULL, p = NULL)
}
\arguments{
\item{o}{(required, numeric vector) Observations, values of a response variable. Default: NULL}

\item{p}{(required, numeric vector) Model predictions. Default: NULL}
}
\value{
numeric: Pearson R-squared
}
\description{
Internal function to compute the R-squared of observations versus model predictions.
}
\seealso{
Other preference_order_tools: 
\code{\link{auc}()},
\code{\link{case_weights}()},
\code{\link{f_auto}()},
\code{\link{f_auto_rules}()},
\code{\link{k_auto}()}
}
\concept{preference_order_tools}
