% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{rescale_vector}
\alias{rescale_vector}
\title{Rescales a numeric vector into a new range}
\usage{
rescale_vector(
  x = NULL,
  new_min = 0,
  new_max = 1,
  old_min = NULL,
  old_max = NULL
)
}
\arguments{
\item{x}{(required, numeric vector) Numeric vector. Default: NULL}

\item{new_min}{(optional, numeric) New minimum value. Default: 0}

\item{new_max}{(optional_numeric) New maximum value. Default: 1}

\item{old_min}{(optional, numeric) Old minimum value. Default: NULL}

\item{old_max}{(optional_numeric) Old maximum value. Default: NULL}
}
\value{
numeric vector with rescaled values of x.
}
\description{
Rescales a numeric vector into a new range
}
\examples{
if(interactive()){

 out <- rescale_vector(
   x = rnorm(100),
   new_min = 0,
   new_max = 100,
   integer = TRUE
   )
   out

}
}
\keyword{internal}
