% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cramer.R
\name{score_cramer}
\alias{score_cramer}
\title{Cramer's V of Observations vs Predictions}
\usage{
score_cramer(o = NULL, p = NULL, ...)
}
\arguments{
\item{o}{(required; character vector) categorical observations. Default: NULL}

\item{p}{(required; character vector) categorical predictions. Default: NULL}

\item{...}{(optional) Used to pass internal arguments such as \code{function_name} for \link{validate_arg_function_name} in nested calls, or \code{m}, a correlation matrix generated via \code{\link[stats:cor]{stats::cor()}} or \code{\link[=cor_matrix]{cor_matrix()}}.}
}
\value{
numeric: Cramer's V
}
\description{
Internal function to compute the Cramer's V of categorical observations versus categorical model predictions. Please read the help file of \code{\link[=cor_cramer]{cor_cramer()}} for further details.
}
\examples{
score_cramer(
 o = c("a", "a", "b", "c", "c"),
 p = c("a", "b", "b", "c", "c")
 )

}
\seealso{
Other modelling_tools: 
\code{\link{case_weights}()},
\code{\link{model_formula}()},
\code{\link{score_auc}()},
\code{\link{score_r2}()}
}
\concept{modelling_tools}
