% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.collinear_output.R
\name{summary.collinear_output}
\alias{summary.collinear_output}
\title{Summary of \code{class.collinear_output}}
\usage{
\method{summary}{collinear_output}(object = NULL, ...)
}
\arguments{
\item{object}{(required, list of class \code{collinear_output}) Object to summarize. Default: NULL}

\item{...}{Ignored, kept for consistency with generic.}
}
\value{
list:
If \code{object} was created with \code{responses = NULL}, a sublist named "result" containing a vector with the selected predictors. Otherwise, a list named after each response containing the corresponding variable selection.
}
\description{
Summary of \code{class.collinear_output}
}
\seealso{
Other S3_methods: 
\code{\link{print.collinear_output}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_selection}()}
}
\concept{S3_methods}
