% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_summary.collinear_output.R
\name{summary.collinear_output}
\alias{summary.collinear_output}
\title{Summary of \code{collinear_output}}
\usage{
\method{summary}{collinear_output}(object = NULL, ...)
}
\arguments{
\item{object}{(required, list of class \code{collinear_output}) Object to summarize. Default: NULL}

\item{...}{Ignored, kept for consistency with generic.}
}
\description{
Summary of \code{collinear_output}
}
