% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_summary.collinear_selection.R
\name{summary.collinear_selection}
\alias{summary.collinear_selection}
\title{Summary Method for Class \code{collinear_selection}}
\usage{
\method{summary}{collinear_selection}(x = NULL)
}
\arguments{
\item{x}{(required, list) Object of class \code{collinear_selection} resulting from \code{\link[=build.collinear_selection]{build.collinear_selection()}}. Default: NULL}
}
\description{
Summary Method for Class \code{collinear_selection}
}
