% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods_summary.collinear_selection.R
\name{summary.collinear_selection}
\alias{summary.collinear_selection}
\title{Summary of \code{class.collinear_selection}}
\usage{
\method{summary}{collinear_selection}(object = NULL, ...)
}
\arguments{
\item{object}{(sub-list in output of \code{\link[=collinear]{collinear()}} ) Object to summarize. Default: NULL}

\item{...}{Ignored, kept for consistency with generic.}
}
\value{
list
TODO describe list structure
}
\description{
Summary of \code{class.collinear_selection}
}
\seealso{
Other S3_methods: 
\code{\link{print.collinear_output}()},
\code{\link{print.collinear_selection}()},
\code{\link{summary.collinear_output}()}
}
\concept{S3_methods}
