% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encode.R
\name{target_encode}
\alias{target_encode}
\title{Target encode a categorical variable based on a numeric variable.}
\usage{
target_encode(numeric = NULL, character = NULL)
}
\arguments{
\item{numeric}{(required; numeric vector) A numeric vector.}

\item{character}{(required; character vector) A character vector of the same length as 'numeric', representing the categorical variable.}
}
\value{
A numeric vector containing target-encoded values.
}
\description{
This function calculates the mean of the numeric variable for each group defined by
the categorical variable and returns the target-encoded values.
}
\examples{
if(interactive()){
data(ecoregions)
encoded_values <- target_encode(
  numeric = ecoregions$plant_richness,
  character = ecoregions$dominant_landcover
  )
}

}
