% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encoder.R
\name{target_encoder}
\alias{target_encoder}
\title{Target encode a categorical variable based on a numeric variable.}
\usage{
target_encoder(x = NULL, y = NULL, check_input = TRUE)
}
\arguments{
\item{x}{(required; numeric vector) A numeric vector generally representing a response variable. Default: NULL}

\item{y}{(required; character vector) A character vector of the same length as 'x', representing the categorical variable.  Default: NULL}

\item{check_input}{(required; logical) If FALSE, disables data checking for a slightly better performance when many tests are performed. Default: TRUE}
}
\value{
A numeric vector containing target-encoded values.
}
\description{
Transforms character variables into numeric by assigning to each group the mean of a numeric variable (usually the model response) across the group's cases.
}
\examples{
if(interactive()){

#load example data frame
data(vi)

biogeo_biome_encoded <- target_encoder(
  x = vi$vi_mean,
  y = vi$biogeo_biome
  )

}

}
