% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{Tolerance for Detecting Linear Dependencies}
\usage{
tolerance()
}
\value{
numeric
}
\description{
Internal function to define the \code{tol} argument in \code{\link[=solve]{solve()}}. Returns \code{0} if the system has long-double capabilities, and \code{.Machine$double.eps} otherwise.
}
\examples{
tolerance()
}
\seealso{
Other data_validation: 
\code{\link{autoconfig_arg_max_cor}()},
\code{\link{autoconfig_arg_max_vif}()},
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_encoding_method}()},
\code{\link{validate_arg_f}()},
\code{\link{validate_arg_function_name}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_quiet}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
