% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_df.R
\name{validate_arg_df}
\alias{validate_arg_df}
\title{Validate Values of Argument \code{df}}
\usage{
validate_arg_df(
  df = NULL,
  response = NULL,
  predictors = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(required; character string or vector) Names of the target response variables in \code{df}. Default: NULL.}

\item{predictors}{(required; character vector) Names of the target predictors in \code{df}. Default: NULL}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console Default: FALSE}
}
\value{
data frame
}
\description{
Internal function to validate the integrity of the argument \code{df}.

It performs the following actions:
\itemize{
\item Returns \code{df} as-is if already tagged with \code{validated = TRUE}.
\item Stops if \code{df} is \code{NULL}, not coercible to a data frame, or has zero rows.
\item Drops geometry column if present.
\item Subsets \code{df} to the \code{response} and \code{predictors} columns.
\item Converts logical columns to numeric.
\item Replaces \code{NaN}, \code{Inf}, and \code{-Inf} values with \code{NA} in numeric columns.
\item Reorders columns to match original \code{response} + \code{predictors} order.
\item Tags the result with the attribute \code{validated = TRUE}.
}
}
\examples{

data(vi, vi_predictors)

vi <- validate_arg_df(
  df = vi,
  response = "vi_numeric",
  predictors = vi_predictors,
  function_name = "f()",
  quiet = FALSE
)

attributes(vi)$validated
}
\seealso{
Other data_validation: 
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
