% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_df_not_null.R
\name{validate_arg_df_not_null}
\alias{validate_arg_df_not_null}
\title{Checks That Argument \code{df} Is Not NULL}
\usage{
validate_arg_df_not_null(df = NULL, function_name = NULL)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}
}
\value{
data frame
}
\description{
Internal function to validate the default value of the argument \code{df}.
}
\examples{
data(vi)
df <- validate_arg_df_not_null(
  df = vi,
  function_name = "f()"
  )
}
\seealso{
Other data_validation: 
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
