% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_encoding_method.R
\name{validate_arg_encoding_method}
\alias{validate_arg_encoding_method}
\title{Validate Argument \code{encoding_method}}
\usage{
validate_arg_encoding_method(
  encoding_method = NULL,
  overwrite = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{encoding_method}{(optional; character or NULL). Name of one target encoding method. One of: "loo", "mean", or "rank" (see \code{\link[=target_encoding_lab]{target_encoding_lab()}} for further details). If NULL, target encoding is disabled. Default: NULL}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}
}
\value{
character
}
\description{
Validate Argument \code{encoding_method}
}
\examples{
x <- validate_arg_encoding_method(
  encoding_method = "wrong_method"
  )
}
