% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_predictors.R
\name{validate_arg_predictors}
\alias{validate_arg_predictors}
\title{Validate Argument \code{predictors}}
\usage{
validate_arg_predictors(
  df = NULL,
  response = NULL,
  predictors = NULL,
  function_name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{response}{(optional; character, character vector, or NULL) Name of one or several response variables in \code{df}, used to target encode categorical predictors, if any, when \code{encoding_method} is not NULL, and to compute preference order when \code{preference_order} is not NULL. If \code{response} is NULL, target encoding and preference order computation are disabled. Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. Default: NULL}

\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console Default: FALSE}
}
\value{
character vector: predictor names
}
\description{
Internal function. Validates the argument \code{predictors} as follows:
\itemize{
\item Returns \code{predictors} as-is if it has the attribute \code{validated = TRUE}.
\item Stops if \code{df} is \code{NULL}, using \code{\link[=validate_arg_df_not_null]{validate_arg_df_not_null()}}.
\item If \code{predictors} is \code{NULL}, uses all column names in \code{df} except those in the \code{response} argument.
\item Removes any elements in \code{predictors} that are also in \code{response}.
\item Ignores \code{predictors} not present as column names in \code{df}, issuing a message if \code{quiet = FALSE}.
\item If \code{df} has at least 10 rows:
\itemize{
\item Removes numeric predictors with near zero variance, with message if \code{quiet = FALSE}.
\item Removes categorical predictors with constant values, with message if \code{quiet = FALSE}.
}
\item Adds the attribute \code{validated = TRUE} to the returned vector.
}
}
\examples{
data(vi, vi_predictors)

predictors <- validate_arg_predictors(
  df = vi,
  predictors = vi_predictors
)

attributes(predictors)$validated

}
\seealso{
Other data_validation: 
\code{\link{validate_arg_df}()},
\code{\link{validate_arg_df_not_null}()},
\code{\link{validate_arg_max_cor}()},
\code{\link{validate_arg_max_vif}()},
\code{\link{validate_arg_predictors_cor}()},
\code{\link{validate_arg_predictors_vif}()},
\code{\link{validate_arg_preference_order}()},
\code{\link{validate_arg_response}()},
\code{\link{validate_args_target_encoding_lab}()}
}
\concept{data_validation}
