% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arg_quiet.R
\name{validate_arg_quiet}
\alias{validate_arg_quiet}
\title{Validate Argument \code{quiet}}
\usage{
validate_arg_quiet(function_name = NULL, quiet = NULL)
}
\arguments{
\item{function_name}{(optional, character string) Name of the function performing the argument check. Default: NULL}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console Default: FALSE}
}
\value{
logical
}
\description{
Internal function to validate the argument \code{quiet}.
}
\examples{

quiet <- validate_arg_quiet(
  function_name = "f()",
  quiet = TRUE
  )

attributes(quiet)$validated

}
