% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_df}
\alias{validate_df}
\title{Validate Argument df}
\usage{
validate_df(df = NULL, min_rows = 30)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{min_rows}{(required; integer) Minimum number of rows required for a pairwise correlation or a variance inflation factor analysis. Default: 30}
}
\value{
data frame
}
\description{
Internal function to validate the argument \code{df} and ensure it complies with the requirements of the package functions. It performs the following actions:
\itemize{
\item Stops if 'df' is NULL.
\item Stops if 'df' cannot be coerced to data frame.
\item Stops if 'df' has zero rows.
\item Removes geometry column if the input data frame is an "sf" object.
\item Removes non-numeric columns with as many unique values as rows df has.
\item Print messages if number of rows of 'df' is lower than 'min_rows'.
\item Converts logical columns to numeric.
\item Converts factor and ordered columns to character.
\item Tags the data frame with the attribute \code{validated = TRUE} to let the package functions skip the data validation.
}
}
\examples{

data(vi)

#validating example data frame
vi <- validate_df(
  df = vi
)

#tagged as validated
attributes(vi)$validated
}
\seealso{
Other data_preparation: 
\code{\link{drop_geometry_column}()},
\code{\link{identify_non_numeric_predictors}()},
\code{\link{identify_numeric_predictors}()},
\code{\link{identify_zero_variance_predictors}()},
\code{\link{validate_predictors}()},
\code{\link{validate_preference_order}()},
\code{\link{validate_response}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_preparation}
