% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_response}
\alias{validate_response}
\title{Validate Argument response}
\usage{
validate_response(df = NULL, response = NULL, decimals = 4)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with numeric predictors, and optionally a numeric response and categorical predictors. Default: NULL.}

\item{response}{(optional, character string) Name of a numeric (non-numerics are ignored) response variable in \code{df}, only required if there are categorical variables named in the argument \code{predictors}. Default: NULL.}

\item{decimals}{(required, integer) Number of decimal places for the zero variance test. Smaller numbers will increase the number of variables detected as near-zero variance. Recommended values will depend on the range of the numeric variables in 'df'. Default: 4}
}
\value{
character string: response name
}
\description{
Internal function to validate the argument \code{response}. Requires the argument 'df' to be validated with \code{\link[=validate_df]{validate_df()}}.
}
\examples{

data(
  vi
)

#validating example data frame
vi <- validate_df(
  df = vi
)

#validating example predictors
response <- validate_response(
  df = vi,
  response = "vi_mean"
)

#tagged as validated
attributes(response)$validated

}
\seealso{
Other data_preparation: 
\code{\link{drop_geometry_column}()},
\code{\link{identify_non_numeric_predictors}()},
\code{\link{identify_zero_variance_predictors}()},
\code{\link{validate_df}()},
\code{\link{validate_predictors}()},
\code{\link{validate_preference_order}()}
}
\author{
Blas M. Benito, PhD
}
\concept{data_preparation}
