% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi}
\alias{vi}
\title{30.000 records of responses and predictors all over the world}
\format{
Data frame with 30.000 rows and 67 columns.
}
\usage{
data(vi)
}
\description{
30.000 records of responses and predictors all over the world
}
\seealso{
\link{vi_predictors}
}
\keyword{datasets}
