% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi_predictors}
\alias{vi_predictors}
\title{Predictor names in data frame 'vi'}
\format{
Character vector with predictor names.
}
\usage{
data(vi_predictors)
}
\description{
Predictor names in data frame 'vi'
}
\seealso{
\link{vi}
}
\keyword{datasets}
