% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vi_responses}
\alias{vi_responses}
\title{Vector of response names in \code{vi} and \code{vi_smol}}
\format{
Character vector with response names.
}
\usage{
data(vi_responses)
}
\description{
Vector of response names in \code{vi} and \code{vi_smol}
}
\seealso{
\link{vi}

Other example_data: 
\code{\link{toy}},
\code{\link{vi}},
\code{\link{vi_predictors}},
\code{\link{vi_predictors_categorical}},
\code{\link{vi_predictors_numeric}},
\code{\link{vi_smol}}
}
\concept{example_data}
\keyword{datasets}
