% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{Variance Inflation Factor from Correlation Matrix}
\usage{
vif(m = NULL)
}
\arguments{
\item{m}{(required, numeric matrix) Correlation matrix. Default: NULL}
}
\value{
data frame: variance inflation factors
}
\description{
Internal function used within \code{\link[=vif_df]{vif_df()}} to compute Variance Inflation Factors from a correlation matrix. Direct usage of this function is not recommended.
}
\examples{
data(
  vi,
  vi_predictors_numeric
)

#compute correlation matrix
m <- stats::cor(
  x = vi[, vi_predictors_numeric[1:10]],
  use = "complete.obs"
)

#compute VIF
vif(m = m)
}
