% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{Variance Inflation Factors From Correlation Matrix}
\usage{
vif(m = NULL)
}
\arguments{
\item{m}{(required, matrix) Correlation matrix generated via \code{\link[stats:cor]{stats::cor()}} or \code{\link[=cor_matrix]{cor_matrix()}}. Must have named dimensions. Default: NULL}
}
\value{
named numeric vector
}
\description{
Computes the Variance Inflation Factors of numeric variables from their correlation matrix in two steps:
\itemize{
\item Applies \code{\link[base:solve]{base::solve()}} to transform the correlation matrix into a precision matrix, which is the inverse of the covariance matrix between all variables in \code{predictors}.
\item Applies \code{\link[base:diag]{base::diag()}} to extract the diagonal of the precision matrix, which contains the variance of the regression of each predictor against all other predictors, also known as Variance Inflation Factor
}
}
\section{Variance Inflation Factors}{


The Variance Inflation Factor for a given variable \eqn{a} is computed as \eqn{1/(1-R2)}, where \eqn{R2} is the multiple R-squared of a multiple regression model fitted using \eqn{a} as response and all other predictors in the input data frame as predictors, as in  \eqn{a = b + c + ...}.

The square root of the VIF of \eqn{a} is the factor by which the confidence interval of the estimate for \eqn{a} in the linear model \eqn{y = a + b + c + ...} is widened by multicollinearity in the model predictors.

The range of VIF values is (1, Inf]. The recommended thresholds for maximum VIF may vary depending on the source consulted, being the most common values, 2.5, 5, and 10.
}

\examples{
data(vi, vi_predictors_numeric)

m <- cor_matrix(
  df = vi[1:1000, ],
  predictors = vi_predictors_numeric[1:5]
)

vif(m)
}
\references{
\itemize{
\item David A. Belsley, D.A., Kuh, E., Welsch, R.E. (1980). Regression Diagnostics: Identifying Influential Data and Sources of Collinearity. John Wiley & Sons. DOI: 10.1002/0471725153.
}
}
\seealso{
Other vif: 
\code{\link{vif_df}()},
\code{\link{vif_select}()}
}
\concept{vif}
