% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_select.R
\name{vif_select}
\alias{vif_select}
\title{Automated Multicollinearity Filtering with Variance Inflation Factors}
\usage{
vif_select(
  df = NULL,
  predictors = NULL,
  preference_order = NULL,
  max_vif = 5,
  quiet = FALSE
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors to select from \code{df}. If omitted, all numeric columns in \code{df} are used instead. Default: NULL}

\item{preference_order}{(optional; character vector, output of \code{\link[=preference_order]{preference_order()}}, list, or NULL). Prioritizes predictors to preserve the most relevant ones during multicollinearity filtering. Accepted inputs are:
\itemize{
\item \strong{NULL} (default): If argument \code{f} is NULL (default), predictors are prioritized from lower to higher multicollinearity. Otherwise it calls \code{\link[=preference_order]{preference_order()}} to rank the predictors using the function defined in the argument \code{f}. The output of this setting might differ to an external call to \code{\link[=preference_order]{preference_order()}} if target encoding is triggered to transform ategorical predictors to numeric.
\item \strong{character vector}: Predictor names in a user-defined priority order. The first predictor in this vector is always selected, unless it has near zero-variance values.
\item \strong{data frame}: output of \code{\link[=preference_order]{preference_order()}} computed on the value of the argument \code{response}.
\item \strong{named list}: list of data frames, output of \code{\link[=preference_order]{preference_order()}} when argument \code{response} is a vector of length two or more.
}. Default: NULL}

\item{max_vif}{(optional, numeric or NULL) Maximum Variance Inflation Factor allowed during variable selection. Recommended values are between 2 (strict) and 10 (permissive). Higher values return larger number of predictors with a higher multicollinearity. If NULL, the variance inflation analysis is disabled. Default: 5.}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console Default: FALSE}
}
\value{
List:
\itemize{
\item timestamp: value of \code{\link[=Sys.time]{Sys.time()}}.
\item arguments: list with input arguments.
\item output: one object of class \code{collinear_selection} generated by \code{\link[=build.collinear_selection]{build.collinear_selection()}} per response in argument \code{response}.
}
}
\description{
This function automatizes multicollinearity filtering in data frames with numeric predictors by combining two methods:
\itemize{
\item \strong{Preference Order}: method to rank and preserve relevant variables during  multicollinearity filtering. See argument \code{preference_order} and function \code{\link[=preference_order]{preference_order()}}.
\item \strong{VIF-based filtering}: recursive algorithm to identify and remove predictors with a VIF above a given threshold.
}

When the argument \code{preference_order} is not provided, the predictors are ranked lower to higher VIF. The predictor selection resulting from this option, albeit diverse and uncorrelated, might not be the one with the highest overall predictive power when used in a model.

Please check the sections \strong{Preference Order}, \strong{Variance Inflation Factors}, and \strong{VIF-based Filtering} at the end of this help file for further details.
}
\section{Preference Order}{


This feature is designed to help protect important predictors during the multicollinearity filtering. It involves the arguments \code{preference_order} and \code{f}.

The argument \code{preference_order} accepts:
\itemize{
\item: A character vector of predictor names in a custom order of preference, from first to last. This vector does not need to contain all predictor names, but only the ones relevant to the user.
\item A data frame returned by \code{\link[=preference_order]{preference_order()}}, which ranks predictors based on their association with a response variable.
\item If NULL, and \code{response} is provided, then \code{\link[=preference_order]{preference_order()}} is used internally to rank the predictors using the function \code{f}. If \code{f} is NULL as well, then \code{\link[=f_auto]{f_auto()}} selects a proper function based on the data properties.
}
}

\section{Variance Inflation Factors}{


The Variance Inflation Factor for a given variable \eqn{a} is computed as \eqn{1/(1-R2)}, where \eqn{R2} is the multiple R-squared of a multiple regression model fitted using \eqn{a} as response and all other predictors in the input data frame as predictors, as in  \eqn{a = b + c + ...}.

The square root of the VIF of \eqn{a} is the factor by which the confidence interval of the estimate for \eqn{a} in the linear model \eqn{y = a + b + c + ...} is widened by multicollinearity in the model predictors.

The range of VIF values is (1, Inf]. The recommended thresholds for maximum VIF may vary depending on the source consulted, being the most common values, 2.5, 5, and 10.
}

\section{VIF-based Filtering}{


The function \code{\link[=vif_select]{vif_select()}} computes Variance Inflation Factors and removes variables iteratively, until all variables in the resulting selection have a VIF below \code{max_vif}.

If the argument \code{preference_order} is not provided, all variables are ranked from lower to higher VIF, as returned by \code{\link[=vif_df]{vif_df()}}, and the variable with the higher VIF above \code{max_vif} is removed on each iteration.

If \code{preference_order} is defined, whenever two or more variables are above \code{max_vif}, the one with a smaller index in the vector \code{preference_order} is preserved, and the next one with a higher VIF is removed. For example, for the predictors and preference order \eqn{a} and \eqn{b}, if any of their VIFs is higher than \code{max_vif}, then \eqn{b} will be removed regardless of whether its VIF is lower or higher than \eqn{a}'s VIF. If their VIF scores are lower than \code{max_vif}, then both are preserved.
}

\examples{
  data(vi, vi_predictors)

  df <- vi[1:1000, ]

  #predictors has mixed types
  sapply(
    X = df[, vi_predictors, drop = FALSE],
    FUN = class
  ) |>
    unique()

  #categorical predictors are ignored
  selected_predictors <- vif_select(
    df = df,
    predictors = vi_predictors,
    max_vif = 5,
    quiet = FALSE
  )

  #all these have a VIF lower than max_vif (2.5)
  vif_df(
    df = df,
    predictors = selected_predictors,
    quiet = TRUE
  )

  #custom preference order
  selected_predictors <- vif_select(
    df = df,
    predictors = vi_predictors,
    preference_order = c(
      "swi_mean",
      "soil_temperature_mean",
      "topo_elevation",
      "wrong_name" #ignored
    ),
    max_vif = 2.5,
    quiet = FALSE
  )

  #using automated preference order
  df_preference <- preference_order(
    df = df,
    response = "vi_numeric",
    predictors = vi_predictors[1:10]
  )

  selected_predictors <- vif_select(
    df = df,
    predictors = vi_predictors,
    preference_order = df_preference,
    max_vif = 5,
    quiet = FALSE
  )
}
\references{
\itemize{
\item David A. Belsley, D.A., Kuh, E., Welsch, R.E. (1980). Regression Diagnostics: Identifying Influential Data and Sources of Collinearity. John Wiley & Sons. DOI: 10.1002/0471725153.
}
}
\seealso{
Other vif: 
\code{\link{vif}()},
\code{\link{vif_df}()}
}
\author{
Blas M. Benito, PhD
}
\concept{vif}
