% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_select.R
\name{vif_select}
\alias{vif_select}
\title{Automated multicollinearity reduction via Variance Inflation Factor}
\usage{
vif_select(
  df = NULL,
  response = NULL,
  predictors = NULL,
  preference_order = NULL,
  max_vif = 5
)
}
\arguments{
\item{df}{(required; data frame or tibble) A data frame with numeric and/or character predictors predictors, and optionally, a response variable. Default: NULL.}

\item{response}{(recommended, character string) Name of a numeric response variable. Character response variables are ignored. Please, see 'Details' to better understand how providing this argument or not leads to different results when there are character variables in 'predictors'. Default: NULL.}

\item{predictors}{(optional; character vector) Character vector with column names of predictors in 'df'. If omitted, all columns of 'df' are used as predictors. Default:'NULL'}

\item{preference_order}{(optional; character vector) vector with column names of 'df' in the desired preference order. Predictors not included in this argument are ranked by their Variance Inflation Factor. Default: \code{NULL}.}

\item{max_vif}{(optional, numeric) Numeric with recommended values between 2.5 and 10 defining the maximum VIF allowed in the output dataset. Higher VIF thresholds should result in a higher number of selected variables. Default: \code{5}.}
}
\value{
Character vector with the names of uncorrelated predictors.
}
\description{
The Variance Inflation Factor for a given variable \code{y} is computed as \code{1/(1-R2)}, where \code{R2} is the multiple R-squared of a multiple regression model fitted using \code{y} as response and all other predictors in the input data frame as predictors. The VIF equation can be interpreted as the "rate of perfect model's R-squared to the unexplained variance of this model".

The possible range of VIF values is (1, Inf]. A VIF lower than 10 suggest that removing \code{y} from the data set would reduce overall multicollinearity. The recommended thresholds for maximum VIF may vary depending on the source consulted, being the most common values, 2.5, 5, and 10.

The function \code{vif_select()} applies a recursive algorithm to remove variables with a VIF higher than a given threshold (defined by the argument \code{max_vif}).

If the argument \code{response} is provided, all non-numeric variables in \code{predictors} are transformed into numeric using target encoding (see \code{target_encode()}). Otherwise, non-numeric variables are ignored.

The argument \code{preference_order} allows defining a preference selection order to preserve (when possible) variables that might be interesting or even required for a given analysis.

For example, if \code{predictors} is \verb{"c("a", "b", "c")} and \code{preference_order} is \verb{"c("a", "b")}, there are two possibilities:
\itemize{
\item If the VIF of \code{"a"} is higher than the VIF of \code{"b"}, and both VIF values are above \code{max_vif}, then \code{"a"} is selected and \code{"b"} is removed.
\item If their correlation is equal or above \code{max_cor}, then \code{"a"} is selected, no matter its correlation with \code{"c"},
}

If \code{preference_order} is not provided, then the predictors are ranked by their variance inflation factor as computed by \code{vif_df()}.
}
\examples{
if(interactive()){

 #loading example data
 data(
  vi,
  vi_predictors
  )

 #without preference order
 selected.variables <- vif_select(
   data = vi,
   predictors = vi_predictors
 )

 selected.variables

 #with preference order
 selected.variables <- vif_select(
   data = vi,
   predictors = vi_predictors,
   preference_order = vi_predictors[1:5],
 )

 selected.variables

}
}
