% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_select.R
\name{vif_select}
\alias{vif_select}
\title{Automated Multicollinearity Filtering with Variance Inflation Factors}
\usage{
vif_select(
  df = NULL,
  predictors = NULL,
  preference_order = NULL,
  max_vif = 5,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{df}{(required; data frame, tibble, or sf) A data frame with responses (optional) and predictors. Must have at least 10 rows for pairwise correlation analysis, and \code{10 * (length(predictors) - 1)} for VIF analysis.  Default: NULL.}

\item{predictors}{(optional; character vector or NULL) Names of the predictors in \code{df} involved in the multicollinearity filtering. If NULL, all columns in \code{df} (except those with constant values or near zero variance) are used. Default: NULL}

\item{preference_order}{(optional; character vector, output of \code{\link[=preference_order]{preference_order()}}, or NULL). Incompatible with \code{f} (overrides it when provided). Prioritizes predictors to preserve the most relevant ones during multicollinearity filtering.

Accepted inputs are:

\itemize{

\item \strong{NULL} (default): If argument \code{f} is NULL (default), predictors are ranked from lower to higher multicollinearity. Otherwise, \code{\link[=preference_order]{preference_order()}} ranks the predictors according to their relationship with \code{responses} using the function defined in \code{f}. NOTE: The output of this setting might differ to an external call to \code{\link[=preference_order]{preference_order()}} if target encoding is triggered.

\item \strong{character vector}: Predictor names in a user-defined priority order. The first predictor in this vector is always selected, unless it has near zero-variance values. This option sets \code{f} to NULL.

\item \strong{data frame}: output of \code{\link[=preference_order]{preference_order()}} computed on the given \code{responses}. This option sets \code{f} to NULL.

\item \strong{named list}: list of data frames, output of \code{\link[=preference_order]{preference_order()}} when argument \code{responses} is a vector of length two or more. This option sets \code{f} to NULL.
}. Default: NULL}

\item{max_vif}{(optional, numeric or NULL) Maximum Variance Inflation Factor allowed for \code{predictors} during multicollinearity filtering. Recommended values are between 2.5 (strict) and 10 (permissive). If NULL, the variance inflation analysis is disabled. Default: 5.}

\item{quiet}{(optional; logical) If FALSE, messages are printed to the console. Default: FALSE}

\item{...}{(optional) Other arguments. Currently used for the internal argument \code{function_name}.}
}
\value{
list of class \code{class.collinear_output}
}
\description{
This function automatizes multicollinearity filtering in data frames with numeric predictors by combining two methods:
\itemize{

\item \strong{Preference Order}: System to rank predictors and protect important ones during multicollinearity filtering. The function offers two alternative options:

\itemize{

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\item Argument \code{preference_order}: Helps preserve predictors of interest for the user. Requires a character vector of predictors names. When two predictors in this vector are highly correlated, the one with a higher index is removed.
}\if{html}{\out{</div>}}

\item If \code{preference_order} is NULL, predictors are ranked from lower to higher VIF, as returned by \code{\link[=vif_df]{vif_df()}}. This option preserves rare predictors over redundant ones, but does not guarantee strong statistical models.

}

\item \strong{VIF-based Filtering}: Computes Variance Inflation Factors for numeric predictors and removes redundant ones iteratively, while taking preference order into account. See \code{\link[=vif]{vif()}} and \code{\link[=vif_df]{vif_df()}} for further details.
}
}
\section{Variance Inflation Factors}{


The Variance Inflation Factor of a given predictor \eqn{a} is computed as \eqn{1/(1-R2)}, where \eqn{R2} is the multiple R-squared of a multiple regression model fitted using \eqn{a} as response against all other predictors, as in  \eqn{a = b + c + ...}.

The square root of the VIF of \eqn{a} is the factor by which the confidence interval of the estimate for \eqn{a} in the linear model \eqn{y = a + b + c + ...} is widened by multicollinearity in the predictors.

The range of VIF values is (1, Inf]. The recommended thresholds for maximum VIF may vary depending on the source consulted, being the most common values, 2.5 (most conservative), 5, and 10 (most permissive).
}

\section{VIF-based Filtering}{


This algorithm, implemented in \code{\link[=vif_select]{vif_select()}}, selects numeric predictors with a Variance Inflation Factor lower than \code{max_vif} while preserving those with a higher ranking in \code{preference_order}, if provided.

The algorithm works as follows:

\enumerate{

\item All numeric predictors are ranked according to the argument \code{preference_order}, or from lower to higher VIF with the other predictors (as computed by \code{\link[=vif_df]{vif_df()}}) otherwise. Categorical predictors are ignored.

\item The predictor with the higher rank is added to the selection.

\item The VIF of the next predictor in the ranking against the predictor/s in the selection is computed via \code{\link[=cor_matrix]{cor_matrix()}} and \code{\link[=vif]{vif()}}. If the resulting VIF is lower than \code{max_vif}, the new predictor is added to the selection. Otherwise it is removed.

\item Step 3 is repeated until all predictors are processed.

}

This filtering method ensures that all predictors in the selection are correlated below \code{max_vif} and that the selection respects the prioritization defined in \code{preference_order}.
}

\examples{
  data(vi, vi_predictors)

  df <- vi[1:1000, ]

  #predictors has mixed types
  sapply(
    X = df[, vi_predictors, drop = FALSE],
    FUN = class
  ) |>
    unique()

  #categorical predictors are ignored
  selected_predictors <- vif_select(
    df = df,
    predictors = vi_predictors,
    max_vif = 5,
    quiet = FALSE
  )

  #all these have a VIF lower than max_vif (2.5)
  vif_df(
    df = df,
    predictors = selected_predictors,
    quiet = TRUE
  )

  #custom preference order
  selected_predictors <- vif_select(
    df = df,
    predictors = vi_predictors,
    preference_order = c(
      "swi_mean",
      "soil_temperature_mean",
      "topo_elevation",
      "wrong_name" #ignored
    ),
    max_vif = 2.5,
    quiet = FALSE
  )

  #using automated preference order
  df_preference <- preference_order(
    df = df,
    responses = "vi_numeric",
    predictors = vi_predictors[1:10]
  )

  selected_predictors <- vif_select(
    df = df,
    predictors = vi_predictors,
    preference_order = df_preference,
    max_vif = 5,
    quiet = FALSE
  )
}
\references{
\itemize{
\item David A. Belsley, D.A., Kuh, E., Welsch, R.E. (1980). Regression Diagnostics: Identifying Influential Data and Sources of Collinearity. John Wiley & Sons. DOI: 10.1002/0471725153.
}
}
\seealso{
Other variance_inflation_factor: 
\code{\link{vif}()},
\code{\link{vif_df}()}
}
\author{
Blas M. Benito, PhD
}
\concept{variance_inflation_factor}
